% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adonis2table.R
\name{adonis2table}
\alias{adonis2table}
\title{FUNCTION_TITLE}
\usage{
adonis2table(mod1, perm_disp = NULL, ancap = NULL, var = "")
}
\arguments{
\item{mod1}{PARAM_DESCRIPTION}

\item{perm_disp}{PARAM_DESCRIPTION, Default: NULL}

\item{ancap}{PARAM_DESCRIPTION, Default: NULL}

\item{var}{PARAM_DESCRIPTION, Default: ''}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
