% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_scImpute.R
\name{calc_scImpute}
\alias{calc_scImpute}
\title{FUNCTION_TITLE}
\usage{
calc_scImpute(otutab)
}
\arguments{
\item{otutab}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
