% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAlphaDiversityTable.R
\name{calculateAlphaDiversityTable}
\alias{calculateAlphaDiversityTable}
\title{calculateAlphaDiversityTable}
\usage{
calculateAlphaDiversityTable(
  phseq_obj,
  outdir,
  indices = c("Observed", "Chao1", "Shannon", "InvSimpson", "Fisher"),
  name = "diversity"
)
}
\arguments{
\item{phseq_obj}{Phyloseq object}

\item{outdir}{Output directory. The results table will be saved here.}

\item{indices}{List of diversity indices to calculate, Default: c("Observed", "Chao1", "Shannon", "InvSimpson", "Fisher")}

\item{name}{PARAM_DESCRIPTION, Default: 'diversity'}
}
\value{
A Data Frame with all the metadata and a variable for each diversity index
}
\description{
Calculates several alpha-diversity indices.
}
\details{
alculates several alpha-diversity indices for each sample in a phyloseq object,
using the phyloseq \code{estimate_richness()} function.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
