% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDoAllModelsFromALLPCAs.R
\name{callDoAllModelsFromALLPCAs}
\alias{callDoAllModelsFromALLPCAs}
\title{FUNCTION_TITLE}
\usage{
callDoAllModelsFromALLPCAs(
  all_pcas,
  name,
  metadata,
  vars2pca = c("Condition"),
  variable_plim = 0.01,
  meta_vars = c(),
  nfolds = 0,
  xgboost_params = xgboost_params,
  catboost_params = catboost_params,
  randomforest_params = randomforest_params,
  do_smote = FALSE,
  smote_params = list(K = 5, dup_size = "balance")
)
}
\arguments{
\item{all_pcas}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}

\item{metadata}{PARAM_DESCRIPTION}

\item{vars2pca}{PARAM_DESCRIPTION, Default: c("Condition")}

\item{variable_plim}{PARAM_DESCRIPTION, Default: 0.01}

\item{meta_vars}{PARAM_DESCRIPTION, Default: c()}

\item{nfolds}{PARAM_DESCRIPTION, Default: 0}

\item{xgboost_params}{PARAM_DESCRIPTION, Default: xgboost_params}

\item{catboost_params}{PARAM_DESCRIPTION, Default: catboost_params}

\item{randomforest_params}{PARAM_DESCRIPTION, Default: randomforest_params}

\item{do_smote}{PARAM_DESCRIPTION, Default: FALSE}

\item{smote_params}{PARAM_DESCRIPTION, Default: list(K = 5, dup_size = "balance")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}}
}
