% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterBrayCurtis.R
\name{clusterBrayCurtis}
\alias{clusterBrayCurtis}
\title{FUNCTION_TITLE}
\usage{
clusterBrayCurtis(
  phobj,
  variable = "Condition",
  renamestr = "",
  dist_type = "bray",
  clust_method = "ward.D2"
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{variable}{PARAM_DESCRIPTION, Default: 'Condition'}

\item{renamestr}{PARAM_DESCRIPTION, Default: ''}

\item{dist_type}{PARAM_DESCRIPTION, Default: 'bray'}

\item{clust_method}{PARAM_DESCRIPTION, Default: 'ward.D2'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[phyloseq]{transform_sample_counts}}, \code{\link[phyloseq]{otu_table}}, \code{\link[phyloseq]{distance}}, \code{\link[phyloseq]{sample_data}}
}
