% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defWriteDEAResults.R
\name{defWriteDEAResults}
\alias{defWriteDEAResults}
\title{FUNCTION_TITLE}
\usage{
defWriteDEAResults(res, resLFC, opt, name)
}
\arguments{
\item{res}{PARAM_DESCRIPTION}

\item{resLFC}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
