% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_full_pipeline.R
\name{deseq_full_pipeline}
\alias{deseq_full_pipeline}
\title{FUNCTION_TITLE}
\usage{
deseq_full_pipeline(phobj, name, vars2deseq, opt)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}

\item{vars2deseq}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{filter}}
}
