% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectOutliers.R
\name{detectOutliers}
\alias{detectOutliers}
\title{FUNCTION_TITLE}
\usage{
detectOutliers(s_meta, vname, p_lim = 0.05)
}
\arguments{
\item{s_meta}{PARAM_DESCRIPTION}

\item{vname}{PARAM_DESCRIPTION}

\item{p_lim}{PARAM_DESCRIPTION, Default: 0.05}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
