% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamLimma_old.R
\name{dreamLimma_old}
\alias{dreamLimma_old}
\title{FUNCTION_TITLE}
\usage{
dreamLimma_old(
  phobj,
  opt,
  variables = c("Condition"),
  individual = "pacienteID",
  outdir = "",
  name = "dream"
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{variables}{PARAM_DESCRIPTION, Default: c("Condition")}

\item{individual}{PARAM_DESCRIPTION, Default: 'pacienteID'}

\item{outdir}{PARAM_DESCRIPTION, Default: ''}

\item{name}{PARAM_DESCRIPTION, Default: 'dream'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[variancePartition]{topTable}}
}
