% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamLimma_old2.R
\name{dreamLimma_old2}
\alias{dreamLimma_old2}
\title{FUNCTION_TITLE}
\usage{
dreamLimma_old2(
  phobj,
  opt,
  variables = c("Condition"),
  individual = "pacienteID"
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{variables}{PARAM_DESCRIPTION, Default: c("Condition")}

\item{individual}{PARAM_DESCRIPTION, Default: 'pacienteID'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
