% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G4Micro-package.R
\docType{package}
\name{G4Micro-package}
\alias{G4Micro}
\alias{G4Micro-package}
\title{G4Micro: Functions to analyse metagenomics data for the Good4Mood depression study (Mora-Martinez, Molina-Mendoza et al.)}
\description{
Functions to analyze microbiome data.
}
\author{
\strong{Maintainer}: Carlos Mora \email{carlos.mora@iata.csic.es}

}
\keyword{internal}
