% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_classif.R
\name{get_classif}
\alias{get_classif}
\title{FUNCTION_TITLE}
\usage{
get_classif(classtring, classnames = NULL)
}
\arguments{
\item{classtring}{PARAM_DESCRIPTION}

\item{classnames}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
