% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signif_components.R
\name{get_signif_components}
\alias{get_signif_components}
\title{FUNCTION_TITLE}
\usage{
get_signif_components(datasc, levs)
}
\arguments{
\item{datasc}{PARAM_DESCRIPTION}

\item{levs}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{arrange}}
}
