% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCazyClass.R
\name{getCazyClass}
\alias{getCazyClass}
\title{FUNCTION_TITLE}
\usage{
getCazyClass(cazy_tt)
}
\arguments{
\item{cazy_tt}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
