% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCLRdata.R
\name{getCLRdata}
\alias{getCLRdata}
\title{FUNCTION_TITLE}
\usage{
getCLRdata(phobj, prevalence_lim = 0.05)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{prevalence_lim}{PARAM_DESCRIPTION, Default: 0.05}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
