% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEdgeColors.R
\name{getEdgeColors}
\alias{getEdgeColors}
\title{FUNCTION_TITLE}
\usage{
getEdgeColors(
  param,
  p.value,
  Estimate,
  plim_plot,
  custom_colors,
  col_ctrl = C_CTRL,
  col_case = C_CASE,
  col_ns = C_NS
)
}
\arguments{
\item{param}{PARAM_DESCRIPTION}

\item{p.value}{PARAM_DESCRIPTION}

\item{Estimate}{PARAM_DESCRIPTION}

\item{plim_plot}{PARAM_DESCRIPTION}

\item{custom_colors}{PARAM_DESCRIPTION}

\item{col_ctrl}{PARAM_DESCRIPTION, Default: C_CTRL}

\item{col_case}{PARAM_DESCRIPTION, Default: C_CASE}

\item{col_ns}{PARAM_DESCRIPTION, Default: C_NS}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
