% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraphFromSamples.R
\name{getGraphFromSamples}
\alias{getGraphFromSamples}
\title{FUNCTION_TITLE}
\usage{
getGraphFromSamples(
  phobj,
  samples,
  daataxa,
  vstdf,
  net_estimator,
  net_method,
  daatab,
  daatab2 = NULL,
  outdir = "./",
  filter_empty = FALSE,
  filt_quantile = 0.95,
  name = "testnet",
  w = 12,
  h = 12
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{samples}{PARAM_DESCRIPTION}

\item{daataxa}{PARAM_DESCRIPTION}

\item{vstdf}{PARAM_DESCRIPTION}

\item{net_estimator}{PARAM_DESCRIPTION}

\item{net_method}{PARAM_DESCRIPTION}

\item{daatab}{PARAM_DESCRIPTION}

\item{daatab2}{PARAM_DESCRIPTION, Default: NULL}

\item{outdir}{PARAM_DESCRIPTION, Default: './'}

\item{filter_empty}{PARAM_DESCRIPTION, Default: FALSE}

\item{filt_quantile}{PARAM_DESCRIPTION, Default: 0.95}

\item{name}{PARAM_DESCRIPTION, Default: 'testnet'}

\item{w}{PARAM_DESCRIPTION, Default: 12}

\item{h}{PARAM_DESCRIPTION, Default: 12}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select_all}}
\code{\link[assertthat]{assert_that}}
}
