% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraphProps.R
\name{getGraphProps}
\alias{getGraphProps}
\title{FUNCTION_TITLE}
\usage{
getGraphProps(net_obj)
}
\arguments{
\item{net_obj}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{arrange}}
}
