% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeanRelAbundancesByGenus.R
\name{getMeanRelAbundancesByGenus}
\alias{getMeanRelAbundancesByGenus}
\title{FUNCTION_TITLE}
\usage{
getMeanRelAbundancesByGenus(phobj)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[phyloseq]{transform_sample_counts}}
}
