% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNullPlot.R
\name{getNullPlot}
\alias{getNullPlot}
\title{FUNCTION_TITLE}
\usage{
getNullPlot(opt, name = "var", error = FALSE)
}
\arguments{
\item{opt}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION, Default: 'var'}

\item{error}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
