% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPropVar_princomp.R
\name{getPropVar_princomp}
\alias{getPropVar_princomp}
\title{FUNCTION_TITLE}
\usage{
getPropVar_princomp(pca)
}
\arguments{
\item{pca}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
