% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRelAbundancesByGenusAndVariable.R
\name{getRelAbundancesByGenusAndVariable}
\alias{getRelAbundancesByGenusAndVariable}
\title{FUNCTION_TITLE}
\usage{
getRelAbundancesByGenusAndVariable(
  phobj,
  variable = "Condition",
  outname = "prevalence_by_genus.tsv",
  oldlevs = c("Control", "Depression")
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{variable}{PARAM_DESCRIPTION, Default: 'Condition'}

\item{outname}{PARAM_DESCRIPTION, Default: 'prevalence_by_genus.tsv'}

\item{oldlevs}{PARAM_DESCRIPTION, Default: c("Control", "Depression")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{summarise}}, \code{\link[dplyr]{mutate}}
\code{\link[tidyr]{gather}}, \code{\link[tidyr]{spread}}
}
