% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTestsForAllCombinations.R
\name{getTestsForAllCombinations}
\alias{getTestsForAllCombinations}
\title{FUNCTION_TITLE}
\usage{
getTestsForAllCombinations(var, group, sep = "-", paired = F)
}
\arguments{
\item{var}{PARAM_DESCRIPTION}

\item{group}{PARAM_DESCRIPTION}

\item{sep}{PARAM_DESCRIPTION, Default: '-'}

\item{paired}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{arrange}}
}
