% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTotalColors.R
\name{getTotalColors}
\alias{getTotalColors}
\title{FUNCTION_TITLE}
\usage{
getTotalColors(
  bacorder,
  plim_plot,
  custom_colors,
  col_ctrl = C_CTRL,
  col_case = C_CASE,
  col_ns = C_NS,
  col_other = C_OTHER
)
}
\arguments{
\item{bacorder}{PARAM_DESCRIPTION}

\item{plim_plot}{PARAM_DESCRIPTION}

\item{custom_colors}{PARAM_DESCRIPTION}

\item{col_ctrl}{PARAM_DESCRIPTION, Default: C_CTRL}

\item{col_case}{PARAM_DESCRIPTION, Default: C_CASE}

\item{col_ns}{PARAM_DESCRIPTION, Default: C_NS}

\item{col_other}{PARAM_DESCRIPTION, Default: C_OTHER}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
