% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValidFactors.R
\name{getValidFactors}
\alias{getValidFactors}
\title{FUNCTION_TITLE}
\usage{
getValidFactors(
  df,
  max_nas = 0.2,
  min_pct = 0.1,
  max_classes = 5,
  exclude = c()
)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{max_nas}{PARAM_DESCRIPTION, Default: 0.2}

\item{min_pct}{PARAM_DESCRIPTION, Default: 0.1}

\item{max_classes}{PARAM_DESCRIPTION, Default: 5}

\item{exclude}{PARAM_DESCRIPTION, Default: c()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
