% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValidNumeric.R
\name{getValidNumeric}
\alias{getValidNumeric}
\title{FUNCTION_TITLE}
\usage{
getValidNumeric(df, max_nas = 0, max_classes = 4)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{max_nas}{PARAM_DESCRIPTION, Default: 0}

\item{max_classes}{PARAM_DESCRIPTION, Default: 4}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
