% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma4functional.R
\name{imma4functional}
\alias{imma4functional}
\title{FUNCTION_TITLE}
\usage{
imma4functional(
  df2,
  metad2,
  interestvar = "Condition",
  covars = c(),
  form = NULL,
  make_all_contrasts = FALSE
)
}
\arguments{
\item{df2}{PARAM_DESCRIPTION}

\item{metad2}{PARAM_DESCRIPTION}

\item{interestvar}{PARAM_DESCRIPTION, Default: 'Condition'}

\item{covars}{PARAM_DESCRIPTION, Default: c()}

\item{form}{PARAM_DESCRIPTION, Default: NULL}

\item{make_all_contrasts}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}
}
