% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_boxplot_nodeprops.R
\name{make_boxplot_nodeprops}
\alias{make_boxplot_nodeprops}
\title{FUNCTION_TITLE}
\usage{
make_boxplot_nodeprops(
  nodeprops,
  outdir,
  name,
  w = 12,
  h = 6,
  signif_levels = c(`***` = 0.001, `**` = 0.01, `*` = 0.05, ns = 1.1),
  correct_pvals = TRUE
)
}
\arguments{
\item{nodeprops}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}

\item{w}{PARAM_DESCRIPTION, Default: 12}

\item{h}{PARAM_DESCRIPTION, Default: 6}

\item{signif_levels}{PARAM_DESCRIPTION, Default: c(\verb{***} = 0.001, \verb{**} = 0.01, \code{*} = 0.05, ns = 1.1)}

\item{correct_pvals}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}
\code{\link[ggsignif]{stat_signif}}
}
