% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_maplot.R
\name{make_maplot}
\alias{make_maplot}
\title{FUNCTION_TITLE}
\usage{
make_maplot(res, opt, name = "maplot.pdf")
}
\arguments{
\item{res}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION, Default: 'maplot.pdf'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[DESeq2]{plotMA}}
}
