% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_volcano.R
\name{make_volcano}
\alias{make_volcano}
\title{FUNCTION_TITLE}
\usage{
make_volcano(res, opt, name, pcol = "pvalue")
}
\arguments{
\item{res}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}

\item{pcol}{PARAM_DESCRIPTION, Default: 'pvalue'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
