% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAllModels.R
\name{makeAllModels}
\alias{makeAllModels}
\title{FUNCTION_TITLE}
\usage{
makeAllModels(
  datasc,
  plim = 0.01,
  opt,
  name = "Condition",
  nfolds = 0,
  xgboost_params = xgboost_params,
  catboost_params = catboost_params,
  randomforest_params = randomforest_params,
  do_smote = FALSE,
  smote_params = smote_params,
  ensemble_param = "BalancedAccuracy_l1out",
  ensemble_minval = 0,
  ensemble_1knn = FALSE,
  do_ensemble_probs = TRUE
)
}
\arguments{
\item{datasc}{PARAM_DESCRIPTION}

\item{plim}{PARAM_DESCRIPTION, Default: 0.01}

\item{opt}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION, Default: 'Condition'}

\item{nfolds}{PARAM_DESCRIPTION, Default: 0}

\item{xgboost_params}{PARAM_DESCRIPTION, Default: xgboost_params}

\item{catboost_params}{PARAM_DESCRIPTION, Default: catboost_params}

\item{randomforest_params}{PARAM_DESCRIPTION, Default: randomforest_params}

\item{do_smote}{PARAM_DESCRIPTION, Default: FALSE}

\item{smote_params}{PARAM_DESCRIPTION, Default: smote_params}

\item{ensemble_param}{PARAM_DESCRIPTION, Default: 'BalancedAccuracy_l1out'}

\item{ensemble_minval}{PARAM_DESCRIPTION, Default: 0}

\item{ensemble_1knn}{PARAM_DESCRIPTION, Default: FALSE}

\item{do_ensemble_probs}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readr]{write_delim}}
}
