% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAllPCoAs.R
\name{makeAllPCoAs}
\alias{makeAllPCoAs}
\title{FUNCTION_TITLE}
\usage{
makeAllPCoAs(
  phobj,
  outdir,
  method = "PCoA",
  name = "PCoAs",
  dist_type = "bray",
  dist_name = "Bray-Curtis",
  vars2plot = c(),
  extradims = 2:5,
  create_pdfs = MULTI_PAGE_PDFS,
  labelsamples = "sampleID",
  w = 12,
  h = 5
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION}

\item{method}{PARAM_DESCRIPTION, Default: 'PCoA'}

\item{name}{PARAM_DESCRIPTION, Default: 'PCoAs'}

\item{dist_type}{PARAM_DESCRIPTION, Default: 'bray'}

\item{dist_name}{PARAM_DESCRIPTION, Default: 'Bray-Curtis'}

\item{vars2plot}{PARAM_DESCRIPTION, Default: c()}

\item{extradims}{PARAM_DESCRIPTION, Default: 2:5}

\item{create_pdfs}{PARAM_DESCRIPTION, Default: MULTI_PAGE_PDFS}

\item{labelsamples}{PARAM_DESCRIPTION, Default: 'sampleID'}

\item{w}{PARAM_DESCRIPTION, Default: 12}

\item{h}{PARAM_DESCRIPTION, Default: 5}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
