% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBarplotDAA3_Int.R
\name{makeBarplotDAA3_Int}
\alias{makeBarplotDAA3_Int}
\title{FUNCTION_TITLE}
\usage{
makeBarplotDAA3_Int(
  daalist,
  outdir,
  plim = 0.05,
  name = "",
  cond_names = list()
)
}
\arguments{
\item{daalist}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION}

\item{plim}{PARAM_DESCRIPTION, Default: 0.05}

\item{name}{PARAM_DESCRIPTION, Default: ''}

\item{cond_names}{PARAM_DESCRIPTION, Default: list()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{filter}}
\code{\link[tidyr]{unite}}
\code{\link[cowplot]{plot_grid}}
}
