% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCCAfromMatrixVegan.R
\name{makeCCAfromMatrixVegan}
\alias{makeCCAfromMatrixVegan}
\title{FUNCTION_TITLE}
\usage{
makeCCAfromMatrixVegan(
  datamat,
  tax_matrix_clr,
  pcvar2retain = 0.9,
  outdir = "",
  name = "CCA_from_CLR"
)
}
\arguments{
\item{datamat}{PARAM_DESCRIPTION}

\item{tax_matrix_clr}{PARAM_DESCRIPTION}

\item{pcvar2retain}{PARAM_DESCRIPTION, Default: 0.9}

\item{outdir}{PARAM_DESCRIPTION, Default: ''}

\item{name}{PARAM_DESCRIPTION, Default: 'CCA_from_CLR'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
