% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeConstrainedOrdinationAll.R
\name{makeConstrainedOrdinationAll}
\alias{makeConstrainedOrdinationAll}
\title{FUNCTION_TITLE}
\usage{
makeConstrainedOrdinationAll(
  phobj,
  opt,
  name = "CCA",
  outdir = "CCA",
  variableList = list(c("Condition")),
  dist_type = "bray"
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION, Default: 'CCA'}

\item{outdir}{PARAM_DESCRIPTION, Default: 'CCA'}

\item{variableList}{PARAM_DESCRIPTION, Default: list(c("Condition"))}

\item{dist_type}{PARAM_DESCRIPTION, Default: 'bray'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
