% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatmapsFromMultipleDeseqResults.R
\name{makeHeatmapsFromMultipleDeseqResults}
\alias{makeHeatmapsFromMultipleDeseqResults}
\title{FUNCTION_TITLE}
\usage{
makeHeatmapsFromMultipleDeseqResults(
  deseq_results_list,
  daa_main,
  main_name,
  vars2plot,
  italics_rownames = T,
  pfilt = 0.05,
  pplot = 0.05,
  name,
  outdir,
  w = 5,
  h = 4
)
}
\arguments{
\item{deseq_results_list}{PARAM_DESCRIPTION}

\item{daa_main}{PARAM_DESCRIPTION}

\item{main_name}{PARAM_DESCRIPTION}

\item{vars2plot}{PARAM_DESCRIPTION}

\item{italics_rownames}{PARAM_DESCRIPTION, Default: T}

\item{pfilt}{PARAM_DESCRIPTION, Default: 0.05}

\item{pplot}{PARAM_DESCRIPTION, Default: 0.05}

\item{name}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION}

\item{w}{PARAM_DESCRIPTION, Default: 5}

\item{h}{PARAM_DESCRIPTION, Default: 4}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}
}
