% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeKmeans.R
\name{makeKmeans}
\alias{makeKmeans}
\title{FUNCTION_TITLE}
\usage{
makeKmeans(datasc, levs, varnames, SEED = 123, folds = c())
}
\arguments{
\item{datasc}{PARAM_DESCRIPTION}

\item{levs}{PARAM_DESCRIPTION}

\item{varnames}{PARAM_DESCRIPTION}

\item{SEED}{PARAM_DESCRIPTION, Default: 123}

\item{folds}{PARAM_DESCRIPTION, Default: c()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}
}
