% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeKnn.R
\name{makeKnn}
\alias{makeKnn}
\title{FUNCTION_TITLE}
\usage{
makeKnn(datasc, levs, nvars, different_ks = c(1, 3, 5, 7, 9, 11, 13))
}
\arguments{
\item{datasc}{PARAM_DESCRIPTION}

\item{levs}{PARAM_DESCRIPTION}

\item{nvars}{PARAM_DESCRIPTION}

\item{different_ks}{PARAM_DESCRIPTION, Default: c(1, 3, 5, 7, 9, 11, 13)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}
}
