% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLinePlotComparingSamePhobjModels.R
\name{makeLinePlotComparingSamePhobjModels}
\alias{makeLinePlotComparingSamePhobjModels}
\title{FUNCTION_TITLE}
\usage{
makeLinePlotComparingSamePhobjModels(
  phname,
  all_model_results,
  opt,
  w = 8,
  h = 12,
  get_pcnames_from = "padj_taxa_res",
  plot_extra = FALSE,
  plot_indiv = TRUE,
  plot_normal_with_smote = FALSE,
  filter_out = c("Ensemble2"),
  order_by_measure = "Accuracy_l1out",
  from_smote = FALSE,
  name = "",
  sel_method_name = "PCA DESeq"
)
}
\arguments{
\item{phname}{PARAM_DESCRIPTION}

\item{all_model_results}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{w}{PARAM_DESCRIPTION, Default: 8}

\item{h}{PARAM_DESCRIPTION, Default: 12}

\item{get_pcnames_from}{PARAM_DESCRIPTION, Default: 'padj_taxa_res'}

\item{plot_extra}{PARAM_DESCRIPTION, Default: FALSE}

\item{plot_indiv}{PARAM_DESCRIPTION, Default: TRUE}

\item{plot_normal_with_smote}{PARAM_DESCRIPTION, Default: FALSE}

\item{filter_out}{PARAM_DESCRIPTION, Default: c("Ensemble2")}

\item{order_by_measure}{PARAM_DESCRIPTION, Default: 'Accuracy_l1out'}

\item{from_smote}{PARAM_DESCRIPTION, Default: FALSE}

\item{name}{PARAM_DESCRIPTION, Default: ''}

\item{sel_method_name}{PARAM_DESCRIPTION, Default: 'PCA DESeq'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{arrange}}
\code{\link[ggpubr]{theme_pubr}}
\code{\link[cowplot]{plot_grid}}
}
