% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLinePlotComparingSamePhobjModels_Cov.R
\name{makeLinePlotComparingSamePhobjModels_Cov}
\alias{makeLinePlotComparingSamePhobjModels_Cov}
\title{FUNCTION_TITLE}
\usage{
makeLinePlotComparingSamePhobjModels_Cov(
  phname,
  condnames,
  all_model_results,
  name,
  opt,
  w = 8,
  h = 12
)
}
\arguments{
\item{phname}{PARAM_DESCRIPTION}

\item{condnames}{PARAM_DESCRIPTION}

\item{all_model_results}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{w}{PARAM_DESCRIPTION, Default: 8}

\item{h}{PARAM_DESCRIPTION, Default: 12}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{summarise}}
\code{\link[ggpubr]{theme_pubr}}
}
