% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLogisticRegressions_pred2.R
\name{makeLogisticRegressions_pred2}
\alias{makeLogisticRegressions_pred2}
\title{FUNCTION_TITLE}
\usage{
makeLogisticRegressions_pred2(depvar, posval, df, vars2test)
}
\arguments{
\item{depvar}{PARAM_DESCRIPTION}

\item{posval}{PARAM_DESCRIPTION}

\item{df}{PARAM_DESCRIPTION}

\item{vars2test}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{glm}}
\code{\link[broom]{reexports}}
\code{\link[dplyr]{desc}}
}
