% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMediationComplex.R
\name{makeMediationComplex}
\alias{makeMediationComplex}
\title{FUNCTION_TITLE}
\usage{
makeMediationComplex(
  df,
  xname = "Edad",
  yname = "Condition_bin",
  medname = "IMC",
  medname2 = "Faecalibacterium_prausnitzii"
)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{xname}{PARAM_DESCRIPTION, Default: 'Edad'}

\item{yname}{PARAM_DESCRIPTION, Default: 'Condition_bin'}

\item{medname}{PARAM_DESCRIPTION, Default: 'IMC'}

\item{medname2}{PARAM_DESCRIPTION, Default: 'Faecalibacterium_prausnitzii'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
