% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMediationComplex_mergedMEdiat2.R
\name{makeMediationComplex_mergedMEdiat2}
\alias{makeMediationComplex_mergedMEdiat2}
\title{FUNCTION_TITLE}
\usage{
makeMediationComplex_mergedMEdiat2(
  df,
  xname = "Edad",
  yname = "Condition_bin",
  medname = "IMC",
  mednames2 = c("Faecalibacterium_prausnitzii")
)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{xname}{PARAM_DESCRIPTION, Default: 'Edad'}

\item{yname}{PARAM_DESCRIPTION, Default: 'Condition_bin'}

\item{medname}{PARAM_DESCRIPTION, Default: 'IMC'}

\item{mednames2}{PARAM_DESCRIPTION, Default: c("Faecalibacterium_prausnitzii")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_starts}}
}
