% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMediationSimple.R
\name{makeMediationSimple}
\alias{makeMediationSimple}
\title{FUNCTION_TITLE}
\usage{
makeMediationSimple(df, xname, yname, medname)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{xname}{PARAM_DESCRIPTION}

\item{yname}{PARAM_DESCRIPTION}

\item{medname}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
