% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMediationSimple_mergedMediat.R
\name{makeMediationSimple_mergedMediat}
\alias{makeMediationSimple_mergedMediat}
\title{FUNCTION_TITLE}
\usage{
makeMediationSimple_mergedMediat(df, x_name, y_name, mednames)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{x_name}{PARAM_DESCRIPTION}

\item{y_name}{PARAM_DESCRIPTION}

\item{mednames}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
