% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMediationSimple_mergedX.R
\name{makeMediationSimple_mergedX}
\alias{makeMediationSimple_mergedX}
\title{FUNCTION_TITLE}
\usage{
makeMediationSimple_mergedX(df, xnames, yname, medname)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{xnames}{PARAM_DESCRIPTION}

\item{yname}{PARAM_DESCRIPTION}

\item{medname}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
