% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMediationSimplePowerCurve.R
\name{makeMediationSimplePowerCurve}
\alias{makeMediationSimplePowerCurve}
\title{FUNCTION_TITLE}
\usage{
makeMediationSimplePowerCurve(
  df,
  xname,
  yname,
  medname,
  param_estimates = c(a = 0.1, b = 0.1, cp = 0.1),
  nboot = 1000,
  nrep = 1000,
  nboot_curve = 100,
  nrep_curve = 100,
  min_n = 100,
  max_n = 500,
  inc_n = 20
)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{xname}{PARAM_DESCRIPTION}

\item{yname}{PARAM_DESCRIPTION}

\item{medname}{PARAM_DESCRIPTION}

\item{param_estimates}{PARAM_DESCRIPTION, Default: c(a = 0.1, b = 0.1, cp = 0.1)}

\item{nboot}{PARAM_DESCRIPTION, Default: 1000}

\item{nrep}{PARAM_DESCRIPTION, Default: 1000}

\item{nboot_curve}{PARAM_DESCRIPTION, Default: 100}

\item{nrep_curve}{PARAM_DESCRIPTION, Default: 100}

\item{min_n}{PARAM_DESCRIPTION, Default: 100}

\item{max_n}{PARAM_DESCRIPTION, Default: 500}

\item{inc_n}{PARAM_DESCRIPTION, Default: 20}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[bmemLavaan]{power.bmem}}
\code{\link[dplyr]{mutate}}
}
