% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMediationSimplePowerCurveLavaan.R
\name{makeMediationSimplePowerCurveLavaan}
\alias{makeMediationSimplePowerCurveLavaan}
\title{FUNCTION_TITLE}
\usage{
makeMediationSimplePowerCurveLavaan(
  df,
  xname,
  yname,
  medname,
  med_res,
  nrep = 1000,
  min_n = 100,
  max_n = 500,
  inc_n = 20,
  error = 0.1,
  ncores = 12
)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{xname}{PARAM_DESCRIPTION}

\item{yname}{PARAM_DESCRIPTION}

\item{medname}{PARAM_DESCRIPTION}

\item{med_res}{PARAM_DESCRIPTION}

\item{nrep}{PARAM_DESCRIPTION, Default: 1000}

\item{min_n}{PARAM_DESCRIPTION, Default: 100}

\item{max_n}{PARAM_DESCRIPTION, Default: 500}

\item{inc_n}{PARAM_DESCRIPTION, Default: 20}

\item{error}{PARAM_DESCRIPTION, Default: 0.1}

\item{ncores}{PARAM_DESCRIPTION, Default: 12}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{summarise}}
}
