% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePCoA.R
\name{makePCoA}
\alias{makePCoA}
\title{FUNCTION_TITLE}
\usage{
makePCoA(
  phobj,
  pcoa.bray,
  evals,
  var2color = "Condition",
  name = "Bray-Curtis",
  extradims = 2:5,
  labelsamples = "sampleID"
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{pcoa.bray}{PARAM_DESCRIPTION}

\item{evals}{PARAM_DESCRIPTION}

\item{var2color}{PARAM_DESCRIPTION, Default: 'Condition'}

\item{name}{PARAM_DESCRIPTION, Default: 'Bray-Curtis'}

\item{extradims}{PARAM_DESCRIPTION, Default: 2:5}

\item{labelsamples}{PARAM_DESCRIPTION, Default: 'sampleID'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[cowplot]{plot_grid}}
}
