% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePermanova.R
\name{makePermanova}
\alias{makePermanova}
\title{FUNCTION_TITLE}
\usage{
makePermanova(
  phobj,
  dist_method = "bray",
  seed = 123,
  exclude_vars = c("sampleID"),
  outname = "permanovas.tsv",
  disp_permutations = 1000
)
}
\arguments{
\item{phobj}{A \code{phyloseq} object containing OTU/ASV table and sample metadata.}

\item{dist_method}{Dissimilarity method to use (e.g., \code{"bray"}, \code{"jaccard"}). Passed to \code{\link[phyloseq]{distance}}. Default: 'bray'}

\item{seed}{Random seed for reproducibility of adonis2. Default: 123.}

\item{exclude_vars}{Character vector of metadata variable names to exclude from PERMANOVA testing. Default: \code{c("sampleID")}.}

\item{outname}{File name to write results table to (.tsv). Default: \code{"permanovas.tsv"}.}

\item{disp_permutations}{Number of permutations to use in PERMANOVA and dispersion tests. Default: \code{1000}.}
}
\value{
A \code{data.frame} with PERMANOVA results, optionally including dispersion, capscale test p-values and adjusted p-values (FDR).
}
\description{
This function performs PERMANOVA, as implemented in the vegan::adonis2 function,
for all the variables included in the metadata of a phyloseq object,
except for those included in the 'exclude_vars' argument. Based on:
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 makePermanova(phobj, dist_method = "bray",
                 seed = 123,
                 exclude_vars = c("sampleID", "patient_name"),
                 outname = "my_output_name")
 }
}
}
\references{
From https://deneflab.github.io/MicrobeMiseq/demos/mothur_2_phyloseq.html#permanova
}
\seealso{
\code{\link[phyloseq]{distance}}, \code{\link[phyloseq]{prune_samples}}
\code{\link[dplyr]{arrange}}
}
