% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePermanovaFormulas.R
\name{makePermanovaFormulas}
\alias{makePermanovaFormulas}
\title{FUNCTION_TITLE}
\usage{
makePermanovaFormulas(
  phobj,
  formulas,
  dist_method = "bray",
  seed = 123,
  outname = "permanovas_mult.tsv"
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{formulas}{PARAM_DESCRIPTION}

\item{dist_method}{PARAM_DESCRIPTION, Default: 'bray'}

\item{seed}{PARAM_DESCRIPTION, Default: 123}

\item{outname}{PARAM_DESCRIPTION, Default: 'permanovas_mult.tsv'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[phyloseq]{distance}}
\code{\link[dplyr]{arrange}}
}
