% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlotBySpecies.R
\name{makePlotBySpecies}
\alias{makePlotBySpecies}
\title{FUNCTION_TITLE}
\usage{
makePlotBySpecies(
  bacnames,
  df_all,
  outdir,
  name,
  quantvar = "IMC_log",
  quantvar_name = "log(IMC)",
  corrmethod = "pearson",
  plim_plot = 0.05,
  w = 8,
  h = 10
)
}
\arguments{
\item{bacnames}{PARAM_DESCRIPTION}

\item{df_all}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}

\item{quantvar}{PARAM_DESCRIPTION, Default: 'IMC_log'}

\item{quantvar_name}{PARAM_DESCRIPTION, Default: 'log(IMC)'}

\item{corrmethod}{PARAM_DESCRIPTION, Default: 'pearson'}

\item{plim_plot}{PARAM_DESCRIPTION, Default: 0.05}

\item{w}{PARAM_DESCRIPTION, Default: 8}

\item{h}{PARAM_DESCRIPTION, Default: 10}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[assertthat]{assert_that}}
\code{\link[dplyr]{select}}, \code{\link[dplyr]{summarise}}
\code{\link[broom]{reexports}}
\code{\link[cowplot]{plot_grid}}
}
