% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlotBySpeciesEffects_BMI.R
\name{makePlotBySpeciesEffects_BMI}
\alias{makePlotBySpeciesEffects_BMI}
\title{FUNCTION_TITLE}
\usage{
makePlotBySpeciesEffects_BMI(
  bacnames,
  df_all,
  res_final,
  outdir,
  name,
  w = 8,
  h = 10,
  plim_plot = 0.05,
  fix_limits = FALSE
)
}
\arguments{
\item{bacnames}{PARAM_DESCRIPTION}

\item{df_all}{PARAM_DESCRIPTION}

\item{res_final}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}

\item{w}{PARAM_DESCRIPTION, Default: 8}

\item{h}{PARAM_DESCRIPTION, Default: 10}

\item{plim_plot}{PARAM_DESCRIPTION, Default: 0.05}

\item{fix_limits}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[assertthat]{assert_that}}
\code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{filter}}
\code{\link[cowplot]{plot_grid}}
}
