% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTopTaxaPlot.R
\name{makeTopTaxaPlot}
\alias{makeTopTaxaPlot}
\title{FUNCTION_TITLE}
\usage{
makeTopTaxaPlot(
  nodeprops,
  ntop,
  outdir,
  name,
  w = 8,
  h = 8,
  manual_scale = c("#A1C6EA", "#FD8B2F", "#A5ABBD")
)
}
\arguments{
\item{nodeprops}{PARAM_DESCRIPTION}

\item{ntop}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}

\item{w}{PARAM_DESCRIPTION, Default: 8}

\item{h}{PARAM_DESCRIPTION, Default: 8}

\item{manual_scale}{PARAM_DESCRIPTION, Default: c("#A1C6EA", "#FD8B2F", "#A5ABBD")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{group_by}}, \code{\link[dplyr]{mutate}}
}
