% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTotalCountPlot.R
\name{makeTotalCountPlot}
\alias{makeTotalCountPlot}
\title{FUNCTION_TITLE}
\usage{
makeTotalCountPlot(metadata, variable, outdir, name = "")
}
\arguments{
\item{metadata}{PARAM_DESCRIPTION}

\item{variable}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION, Default: ''}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
