% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVenn.R
\name{makeVenn}
\alias{makeVenn}
\title{FUNCTION_TITLE}
\usage{
makeVenn(vars2venn, name = "VennDiagram", opt, w = 5, h = 5)
}
\arguments{
\item{vars2venn}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION, Default: 'VennDiagram'}

\item{opt}{PARAM_DESCRIPTION}

\item{w}{PARAM_DESCRIPTION, Default: 5}

\item{h}{PARAM_DESCRIPTION, Default: 5}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
