% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAgainstAllFactors_pred2.R
\name{plotAgainstAllFactors_pred2}
\alias{plotAgainstAllFactors_pred2}
\title{FUNCTION_TITLE}
\usage{
plotAgainstAllFactors_pred2(
  alldf,
  groupvars,
  name,
  outdir,
  opt,
  correct_bonferroni = F,
  names_cyt = c()
)
}
\arguments{
\item{alldf}{PARAM_DESCRIPTION}

\item{groupvars}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{correct_bonferroni}{PARAM_DESCRIPTION, Default: F}

\item{names_cyt}{PARAM_DESCRIPTION, Default: c()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
