% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAgainstAllQuant_pred2.R
\name{plotAgainstAllQuant_pred2}
\alias{plotAgainstAllQuant_pred2}
\title{FUNCTION_TITLE}
\usage{
plotAgainstAllQuant_pred2(
  alldf,
  groupvars,
  name,
  outdir,
  col_var,
  opt,
  names_cyt = c()
)
}
\arguments{
\item{alldf}{PARAM_DESCRIPTION}

\item{groupvars}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION}

\item{col_var}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{names_cyt}{PARAM_DESCRIPTION, Default: c()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tidyr]{gather}}
}
