% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAgeMediationSimple.R
\name{plotAgeMediationSimple}
\alias{plotAgeMediationSimple}
\title{FUNCTION_TITLE}
\usage{
plotAgeMediationSimple(
  res,
  vars2test,
  outdir,
  outname,
  plim_plot = 0.05,
  use_color_scale = FALSE,
  w = 14,
  h = 6
)
}
\arguments{
\item{res}{PARAM_DESCRIPTION}

\item{vars2test}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION}

\item{outname}{PARAM_DESCRIPTION}

\item{plim_plot}{PARAM_DESCRIPTION, Default: 0.05}

\item{use_color_scale}{PARAM_DESCRIPTION, Default: FALSE}

\item{w}{PARAM_DESCRIPTION, Default: 14}

\item{h}{PARAM_DESCRIPTION, Default: 6}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[assertthat]{assert_that}}
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}
\code{\link[scales]{pal_gradient_n}}
}
