% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBoxplots_pred2.R
\name{plotBoxplots_pred2}
\alias{plotBoxplots_pred2}
\title{FUNCTION_TITLE}
\usage{
plotBoxplots_pred2(
  df,
  variable,
  grouping_var,
  wrap_var,
  fname,
  signif_levels = c(`***` = 0.001, `**` = 0.01, `*` = 0.05, ns = 1.1),
  num_comparisons = 1,
  ylabel = "pg/mL",
  correct_pvalues = TRUE,
  write = TRUE
)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{variable}{PARAM_DESCRIPTION}

\item{grouping_var}{PARAM_DESCRIPTION}

\item{wrap_var}{PARAM_DESCRIPTION}

\item{fname}{PARAM_DESCRIPTION}

\item{signif_levels}{PARAM_DESCRIPTION, Default: c(\verb{***} = 0.001, \verb{**} = 0.01, \code{*} = 0.05, ns = 1.1)}

\item{num_comparisons}{PARAM_DESCRIPTION, Default: 1}

\item{ylabel}{PARAM_DESCRIPTION, Default: 'pg/mL'}

\item{correct_pvalues}{PARAM_DESCRIPTION, Default: TRUE}

\item{write}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
