% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsDeseq.R
\name{plotCountsDeseq}
\alias{plotCountsDeseq}
\title{FUNCTION_TITLE}
\usage{
plotCountsDeseq(dds, raw_counts, norm_counts)
}
\arguments{
\item{dds}{PARAM_DESCRIPTION}

\item{raw_counts}{PARAM_DESCRIPTION}

\item{norm_counts}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
