% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetWithTopNames.R
\name{plotNetWithTopNames}
\alias{plotNetWithTopNames}
\title{FUNCTION_TITLE}
\usage{
plotNetWithTopNames(
  net_obj,
  nodeprops,
  net_class,
  ntop = 10,
  outdir,
  name,
  add_labels = TRUE,
  w = 8,
  h = 8
)
}
\arguments{
\item{net_obj}{PARAM_DESCRIPTION}

\item{nodeprops}{PARAM_DESCRIPTION}

\item{net_class}{PARAM_DESCRIPTION}

\item{ntop}{PARAM_DESCRIPTION, Default: 10}

\item{outdir}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}

\item{add_labels}{PARAM_DESCRIPTION, Default: TRUE}

\item{w}{PARAM_DESCRIPTION, Default: 8}

\item{h}{PARAM_DESCRIPTION, Default: 8}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}
}
