% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA_pred2.R
\name{plotPCA_pred2}
\alias{plotPCA_pred2}
\title{FUNCTION_TITLE}
\usage{
plotPCA_pred2(
  dfraw,
  vars2pc,
  labvar = "PACIENTE",
  plotvars = c("Condition"),
  transform = "scale",
  dims = 2:5,
  name = "PCAs.pdf",
  outdir = "",
  w = 12,
  h = 8
)
}
\arguments{
\item{dfraw}{PARAM_DESCRIPTION}

\item{vars2pc}{PARAM_DESCRIPTION}

\item{labvar}{PARAM_DESCRIPTION, Default: 'PACIENTE'}

\item{plotvars}{PARAM_DESCRIPTION, Default: c("Condition")}

\item{transform}{PARAM_DESCRIPTION, Default: 'scale'}

\item{dims}{PARAM_DESCRIPTION, Default: 2:5}

\item{name}{PARAM_DESCRIPTION, Default: 'PCAs.pdf'}

\item{outdir}{PARAM_DESCRIPTION, Default: ''}

\item{w}{PARAM_DESCRIPTION, Default: 12}

\item{h}{PARAM_DESCRIPTION, Default: 8}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[cowplot]{plot_grid}}
}
