% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCsCorwithASVs.R
\name{plotPCsCorwithASVs}
\alias{plotPCsCorwithASVs}
\title{FUNCTION_TITLE}
\usage{
plotPCsCorwithASVs(tab, contrres, num_model = 1, component = "Comp.2")
}
\arguments{
\item{tab}{PARAM_DESCRIPTION}

\item{contrres}{PARAM_DESCRIPTION}

\item{num_model}{PARAM_DESCRIPTION, Default: 1}

\item{component}{PARAM_DESCRIPTION, Default: 'Comp.2'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
