% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCsCorwithASVs_PSO.R
\name{plotPCsCorwithASVs_PSO}
\alias{plotPCsCorwithASVs_PSO}
\title{FUNCTION_TITLE}
\usage{
plotPCsCorwithASVs_PSO(tab, contrres, num_model = 1)
}
\arguments{
\item{tab}{PARAM_DESCRIPTION}

\item{contrres}{PARAM_DESCRIPTION}

\item{num_model}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
