% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhylumBoxplots.R
\name{plotPhylumBoxplots}
\alias{plotPhylumBoxplots}
\title{Plot Boxplots of Phylum-Level Abundances}
\usage{
plotPhylumBoxplots(
  phobj,
  var = "Condition",
  outname = "phylumBarplot.pdf",
  height = 10,
  width = 8,
  paired = F
)
}
\arguments{
\item{phobj}{A phyloseq object containing microbiome data.}

\item{var}{A sample metadata variable used to group the boxplots, Default: 'Condition'}

\item{outname}{Name of the output PDF file, Default: 'phylumBarplot.pdf'}

\item{height}{Height (in inches) of the output PDF, Default: 10}

\item{width}{Width (in inches) of the output PDF, Default: 8}

\item{paired}{Logical indicating if paired Wilcoxon tests should be used, Default: FALSE}
}
\value{
A ggplot2 object representing the boxplots.
}
\description{
Plots boxplots of phylum-level abundances using a phyloseq object, faceted by phylum,
with optional paired comparisons between groups. Significance is assessed using Wilcoxon tests.
}
\details{
This function summarizes abundances at the phylum level, then creates boxplots grouped by the
specified metadata variable. Significance comparisons between groups are performed using
\code{stat_compare_means} from the ggpubr package.
}
\examples{
\dontrun{
if(interactive()){
 plotPhylumBoxplots(my_phyloseq_object)
 }
}
}
\seealso{
\code{\link[phyloseq]{tax_glom}},
\code{\link[phyloseq]{taxa_names}},
\code{\link[phyloseq]{tax_table}},
\code{\link[phyloseq]{psmelt}}
\code{\link[ggpubr]{stat_compare_means}},
\code{\link[ggplot2]{ggplot}},
\code{\link[ggplot2]{geom_boxplot}},
\code{\link[ggplot2]{geom_jitter}},
\code{\link[ggplot2]{geom_line}},
\code{\link[ggplot2]{facet_wrap}}
}
