% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegr_pred2.R
\name{plotRegr_pred2}
\alias{plotRegr_pred2}
\title{FUNCTION_TITLE}
\usage{
plotRegr_pred2(df, variable, x_var, wrap_var, col_var, fname, write = TRUE)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{variable}{PARAM_DESCRIPTION}

\item{x_var}{PARAM_DESCRIPTION}

\item{wrap_var}{PARAM_DESCRIPTION}

\item{col_var}{PARAM_DESCRIPTION}

\item{fname}{PARAM_DESCRIPTION}

\item{write}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[cowplot]{plot_grid}}
}
