% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRelativeAbnBarsPhylum.R
\name{plotRelativeAbnBarsPhylum}
\alias{plotRelativeAbnBarsPhylum}
\title{FUNCTION_TITLE}
\usage{
plotRelativeAbnBarsPhylum(
  phobj,
  variable = "Condition",
  outname = "phylumBarplot.pdf",
  height = 8,
  width = 12,
  ocluster = F
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{variable}{PARAM_DESCRIPTION, Default: 'Condition'}

\item{outname}{PARAM_DESCRIPTION, Default: 'phylumBarplot.pdf'}

\item{height}{PARAM_DESCRIPTION, Default: 8}

\item{width}{PARAM_DESCRIPTION, Default: 12}

\item{ocluster}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[phyloseq]{transform_sample_counts}}, \code{\link[phyloseq]{plot_bar}}
\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{summarise}}
}
