% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_names2.R
\name{process_names2}
\alias{process_names2}
\title{FUNCTION_TITLE}
\usage{
process_names2(namelist)
}
\arguments{
\item{namelist}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
