% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restauraropt_mk.R
\name{restauraropt_mk}
\alias{restauraropt_mk}
\title{FUNCTION_TITLE}
\usage{
restauraropt_mk(opt)
}
\arguments{
\item{opt}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
