% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePsWithLogs.R
\name{updatePsWithLogs}
\alias{updatePsWithLogs}
\title{FUNCTION_TITLE}
\usage{
updatePsWithLogs(phobj, vars = c("nreads"))
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{vars}{PARAM_DESCRIPTION, Default: c("nreads")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
