% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteManyPlots.R
\name{WriteManyPlots}
\alias{WriteManyPlots}
\title{FUNCTION_TITLE}
\usage{
WriteManyPlots(
  plotlist,
  name,
  outdir,
  w = 12,
  h = 7,
  separate = F,
  opt = list()
)
}
\arguments{
\item{plotlist}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION}

\item{w}{PARAM_DESCRIPTION, Default: 12}

\item{h}{PARAM_DESCRIPTION, Default: 7}

\item{separate}{PARAM_DESCRIPTION, Default: F}

\item{opt}{PARAM_DESCRIPTION, Default: list()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
