% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicLimma.R
\name{basicLimma}
\alias{basicLimma}
\title{FUNCTION_TITLE}
\usage{
basicLimma(
  phobj,
  opt,
  variables = c("Condition"),
  individual = "pacienteID",
  outdir = "",
  name = "basic_limma"
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{variables}{PARAM_DESCRIPTION, Default: c("Condition")}

\item{individual}{PARAM_DESCRIPTION, Default: 'pacienteID'}

\item{outdir}{PARAM_DESCRIPTION, Default: ''}

\item{name}{PARAM_DESCRIPTION, Default: 'basic_limma'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
