% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAlphaDiversityTable.R
\name{calculateAlphaDiversityTable}
\alias{calculateAlphaDiversityTable}
\title{FUNCTION_TITLE}
\usage{
calculateAlphaDiversityTable(
  phseq_obj,
  outdir,
  indices = c("Observed", "Chao1", "Shannon", "InvSimpson", "Fisher"),
  name = "diversity"
)
}
\arguments{
\item{phseq_obj}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION}

\item{indices}{PARAM_DESCRIPTION, Default: c("Observed", "Chao1", "Shannon", "InvSimpson", "Fisher")}

\item{name}{PARAM_DESCRIPTION, Default: 'diversity'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
