% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareLFCContrats3.R
\name{compareLFCContrats3}
\alias{compareLFCContrats3}
\title{FUNCTION_TITLE}
\usage{
compareLFCContrats3(
  contrastlist,
  firstContrast,
  contrastNamesOrdered,
  mainContrastName,
  plim_select = 1e-06,
  plim_plot = 0.05,
  name2remove = "",
  resdfname = "resdf",
  outdir = "./",
  name = "LFC_compare3",
  w = 8,
  h = 4,
  scale_mode = "fixed"
)
}
\arguments{
\item{contrastlist}{PARAM_DESCRIPTION}

\item{firstContrast}{PARAM_DESCRIPTION}

\item{contrastNamesOrdered}{PARAM_DESCRIPTION}

\item{mainContrastName}{PARAM_DESCRIPTION}

\item{plim_select}{PARAM_DESCRIPTION, Default: 1e-06}

\item{plim_plot}{PARAM_DESCRIPTION, Default: 0.05}

\item{name2remove}{PARAM_DESCRIPTION, Default: ''}

\item{resdfname}{PARAM_DESCRIPTION, Default: 'resdf'}

\item{outdir}{PARAM_DESCRIPTION, Default: './'}

\item{name}{PARAM_DESCRIPTION, Default: 'LFC_compare3'}

\item{w}{PARAM_DESCRIPTION, Default: 8}

\item{h}{PARAM_DESCRIPTION, Default: 4}

\item{scale_mode}{PARAM_DESCRIPTION, Default: 'fixed'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}
}
