% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defWriteMatAsDF.R
\name{defWriteMatAsDF}
\alias{defWriteMatAsDF}
\title{FUNCTION_TITLE}
\usage{
defWriteMatAsDF(mat, opt, name)
}
\arguments{
\item{mat}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
