% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_full_pipeline.R
\name{deseq_full_pipeline}
\alias{deseq_full_pipeline}
\title{Full DESeq2 Differential Expression Analysis Pipeline}
\usage{
deseq_full_pipeline(
  phobj,
  name,
  vars2deseq,
  opt,
  plot_all = TRUE,
  doPoscounts = FALSE,
  deseqname = "DeSEQ2/",
  vars2heatmap = c("Condition"),
  max_hm_h = 16,
  max_hm_w = 7
)
}
\arguments{
\item{phobj}{A phyloseq object containing count data and taxonomy information.}

\item{name}{A character string used as a prefix for output directories and files.}

\item{vars2deseq}{Character vector of variable names to use in the DESeq2 design formula.}

\item{opt}{A list of options controlling output directories, filtering thresholds, p-value cutoffs, and other parameters.}

\item{plot_all}{Whether to make plots for all contrasts or not. Default, TRUE.}

\item{doPoscounts}{Logical; whether to force the use of the \code{poscounts} normalization method. Default: FALSE.}

\item{deseqname}{A character string specifying the subdirectory name for storing DESeq2 results.
Default: \code{"DeSEQ2/"}.}

\item{vars2heatmap}{Character vector of variable names (from sample metadata) to annotate
in generated heatmaps. Default: \code{c("Condition")}.}

\item{max_hm_h}{Numeric; maximum height of the heatmaps (in inches) when saving PDF output.
Passed to \code{makeHeatmap}. Default: \code{16}.}

\item{max_hm_w}{Max heatmap width. If 0, will be automatically determined depending on number of cols. Default: `7}
}
\value{
A list containing DESeq2 result data frames and associated objects produced by \code{getDeseqResults}.
}
\description{
Runs a comprehensive differential expression analysis pipeline on a phyloseq object,
including filtering, DESeq2 result generation, annotation, visualization (MA plots, volcano plots, heatmaps),
and saving of annotated results.
}
\details{
This function creates output directories if needed, computes the minimum number of samples with counts based on frequency thresholds,
executes differential expression analysis through \code{getDeseqResults()}, annotates results with genus-level taxonomy,
saves annotated tables as TSV files, and generates diagnostic and visualization plots such as MA plots, volcano plots, and heatmaps.
It also performs intersection of taxa lists across contrasts when applicable.
}
\examples{
\dontrun{
if(interactive()){
  opt <- list(out = "results/", minfreq = 0.01, pval = 0.05)
  vars <- c("Condition", "Batch")
  res <- deseq_full_pipeline(physeq_object, "MyExperiment", vars, opt)
}
}
}
\seealso{
\code{\link{getDeseqResults}}, \code{\link{make_maplot}}, \code{\link{make_volcano}},
\code{\link{makeHeatmap}}, \code{\link{getSummaryTablesDeseq}}
\code{\link[DESeq2]{DESeq}}, \code{\link[DESeq2]{results}}, \code{\link[DESeq2]{plotDispEsts}},
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{filter}}
}
