% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxa_padj.R
\name{filter_taxa_padj}
\alias{filter_taxa_padj}
\title{FUNCTION_TITLE}
\usage{
filter_taxa_padj(resdf, plim = 0.05, fc = 1)
}
\arguments{
\item{resdf}{PARAM_DESCRIPTION}

\item{plim}{PARAM_DESCRIPTION, Default: 0.05}

\item{fc}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}
}
