% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterWholeProcessesAndFreq.R
\name{filterWholeProcessesAndFreq}
\alias{filterWholeProcessesAndFreq}
\title{FUNCTION_TITLE}
\usage{
filterWholeProcessesAndFreq(df, opt)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}
}
