% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signif_components_multiclass.R
\name{get_signif_components_multiclass}
\alias{get_signif_components_multiclass}
\title{FUNCTION_TITLE}
\usage{
get_signif_components_multiclass(datasc, levs, plim = 0.05)
}
\arguments{
\item{datasc}{PARAM_DESCRIPTION}

\item{levs}{PARAM_DESCRIPTION}

\item{plim}{PARAM_DESCRIPTION, Default: 0.05}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{arrange}}
\code{\link[broom]{reexports}}
}
