% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlphaDiversity.R
\name{getAlphaDiversity}
\alias{getAlphaDiversity}
\title{FUNCTION_TITLE}
\usage{
getAlphaDiversity(
  phseq_obj,
  vars,
  qvars = c(),
  opt = list(),
  indices = c("Observed", "Chao1", "Shannon", "InvSimpson", "Fisher"),
  name = "AlphaDiversity",
  signif_levels = c(`***` = 0.001, `**` = 0.01, `*` = 0.05, ns = 1.1),
  correct_pvalues = TRUE,
  correct_pvalues_indices = FALSE,
  test2show = "wilcox.test",
  w = 8,
  h = 4
)
}
\arguments{
\item{phseq_obj}{PARAM_DESCRIPTION}

\item{vars}{PARAM_DESCRIPTION}

\item{qvars}{PARAM_DESCRIPTION, Default: c()}

\item{opt}{PARAM_DESCRIPTION, Default: list()}

\item{indices}{PARAM_DESCRIPTION, Default: c("Observed", "Chao1", "Shannon", "InvSimpson", "Fisher")}

\item{name}{PARAM_DESCRIPTION, Default: 'AlphaDiversity'}

\item{signif_levels}{PARAM_DESCRIPTION, Default: c(\verb{***} = 0.001, \verb{**} = 0.01, \code{*} = 0.05, ns = 1.1)}

\item{correct_pvalues}{PARAM_DESCRIPTION, Default: TRUE}

\item{correct_pvalues_indices}{PARAM_DESCRIPTION, Default: FALSE}

\item{test2show}{PARAM_DESCRIPTION, Default: 'wilcox.test'}

\item{w}{PARAM_DESCRIPTION, Default: 8}

\item{h}{PARAM_DESCRIPTION, Default: 4}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[base]{subset}}
\code{\link[phyloseq]{prune_samples}}
\code{\link[ggsignif]{stat_signif}}
\code{\link[dplyr]{mutate}}
}
