% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorrelations.R
\name{getCorrelations}
\alias{getCorrelations}
\title{FUNCTION_TITLE}
\usage{
getCorrelations(mat1, mat2)
}
\arguments{
\item{mat1}{PARAM_DESCRIPTION}

\item{mat2}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
