% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorsByGroup.R
\name{getCorsByGroup}
\alias{getCorsByGroup}
\title{FUNCTION_TITLE}
\usage{
getCorsByGroup(
  mat1,
  mat2,
  group1,
  group2,
  outdir,
  name,
  levnames = c("PSO", "CTRL")
)
}
\arguments{
\item{mat1}{PARAM_DESCRIPTION}

\item{mat2}{PARAM_DESCRIPTION}

\item{group1}{PARAM_DESCRIPTION}

\item{group2}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}

\item{levnames}{PARAM_DESCRIPTION, Default: c("PSO", "CTRL")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}
}
