% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDeseqContrastFromCategorical.R
\name{getDeseqContrastFromCategorical}
\alias{getDeseqContrastFromCategorical}
\title{FUNCTION_TITLE}
\usage{
getDeseqContrastFromCategorical(dds, lev_combin, opt, name)
}
\arguments{
\item{dds}{PARAM_DESCRIPTION}

\item{lev_combin}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
