% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDeseqContrastFromNumerical.R
\name{getDeseqContrastFromNumerical}
\alias{getDeseqContrastFromNumerical}
\title{FUNCTION_TITLE}
\usage{
getDeseqContrastFromNumerical(dds, nvarname, opt, name)
}
\arguments{
\item{dds}{PARAM_DESCRIPTION}

\item{nvarname}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
