% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDeseqResults.R
\name{getDeseqResults}
\alias{getDeseqResults}
\title{FUNCTION_TITLE}
\usage{
getDeseqResults(
  phobj,
  opt,
  name = "",
  variables = c("Condition"),
  formula = NULL,
  doposcounts = FALSE
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION, Default: ''}

\item{variables}{PARAM_DESCRIPTION, Default: c("Condition")}

\item{formula}{PARAM_DESCRIPTION, Default: NULL}

\item{doposcounts}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
