% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDeseqResults_fromFormula.R
\name{getDeseqResults_fromFormula}
\alias{getDeseqResults_fromFormula}
\title{FUNCTION_TITLE}
\usage{
getDeseqResults_fromFormula(phobj, opt, name = "", formula = "~ Condition")
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION, Default: ''}

\item{formula}{PARAM_DESCRIPTION, Default: '~ Condition'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
