% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDeseqResults_paired.R
\name{getDeseqResults_paired}
\alias{getDeseqResults_paired}
\title{FUNCTION_TITLE}
\usage{
getDeseqResults_paired(
  phobj,
  opt,
  name = "",
  variables = c("Condition"),
  individual = "pacienteID"
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION, Default: ''}

\item{variables}{PARAM_DESCRIPTION, Default: c("Condition")}

\item{individual}{PARAM_DESCRIPTION, Default: 'pacienteID'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
