% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGTTableFromRes.R
\name{getGTTableFromRes}
\alias{getGTTableFromRes}
\title{FUNCTION_TITLE}
\usage{
getGTTableFromRes(res, genes, name)
}
\arguments{
\item{res}{PARAM_DESCRIPTION}

\item{genes}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
