% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeanRelAbundancesByPhylum.R
\name{getMeanRelAbundancesByPhylum}
\alias{getMeanRelAbundancesByPhylum}
\title{FUNCTION_TITLE}
\usage{
getMeanRelAbundancesByPhylum(phobj)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
