% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCnamesFromAllresults.R
\name{getPCnamesFromAllresults}
\alias{getPCnamesFromAllresults}
\title{FUNCTION_TITLE}
\usage{
getPCnamesFromAllresults(
  phname,
  all_model_results,
  get_pcnames_from = "padj_taxa_res",
  pca_name = "padj_taxa_pcas",
  varname = "Condition"
)
}
\arguments{
\item{phname}{PARAM_DESCRIPTION}

\item{all_model_results}{PARAM_DESCRIPTION}

\item{get_pcnames_from}{PARAM_DESCRIPTION, Default: 'padj_taxa_res'}

\item{pca_name}{PARAM_DESCRIPTION, Default: 'padj_taxa_pcas'}

\item{varname}{PARAM_DESCRIPTION, Default: 'Condition'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
