% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhylumTests.R
\name{getPhylumTests}
\alias{getPhylumTests}
\title{FUNCTION_TITLE}
\usage{
getPhylumTests(phobj, var = "Condition", outname = "phylumTests", paired = F)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{var}{PARAM_DESCRIPTION, Default: 'Condition'}

\item{outname}{PARAM_DESCRIPTION, Default: 'phylumTests'}

\item{paired}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[phyloseq]{tax_glom}}, \code{\link[phyloseq]{taxa_names}}, \code{\link[phyloseq]{tax_table}}, \code{\link[phyloseq]{psmelt}}
\code{\link[dplyr]{group_map}}
\code{\link[broom]{reexports}}
}
