% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPropVar.R
\name{getPropVar}
\alias{getPropVar}
\title{FUNCTION_TITLE}
\usage{
getPropVar(pca, PC)
}
\arguments{
\item{pca}{PARAM_DESCRIPTION}

\item{PC}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
