% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPropVar_princomp.R
\name{getPropVar_princomp}
\alias{getPropVar_princomp}
\title{Generate Labels of Variance Explained from a princomp Object}
\usage{
getPropVar_princomp(pca)
}
\arguments{
\item{pca}{A \code{princomp} class object resulting from PCA analysis.}
}
\value{
A named character vector with labels like "PC1 (xx.xx\%)" representing variance explained by each component.
}
\description{
Creates character labels for each principal component indicating the percentage of variance explained, based on a \code{princomp} PCA object.
}
\details{
The function calculates the proportion of variance explained by each principal component, rounds it to two decimals, and formats the result into readable labels for visualization or reporting.
}
\examples{
\dontrun{
if(interactive()){
  data(USArrests)
  pca <- princomp(USArrests, cor=TRUE)
  getPropVar_princomp(pca)
}
}
}
