% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRelAbundancesByPhylumAndVariable.R
\name{getRelAbundancesByPhylumAndVariable}
\alias{getRelAbundancesByPhylumAndVariable}
\title{Phylum-Level Abundance and Prevalence Summary by Group}
\usage{
getRelAbundancesByPhylumAndVariable(
  phobj,
  variable = "Condition",
  outname = "prevalence_by_phylum.tsv",
  oldlevs = c("Control", "Depression")
)
}
\arguments{
\item{phobj}{A \code{phyloseq} object containing microbiome data.}

\item{variable}{A character string indicating the sample metadata variable to group by. Default: 'Condition'}

\item{outname}{Filename to save the output summary table. Set to "" to avoid saving. Default: 'prevalence_by_phylum.tsv'}

\item{oldlevs}{A character vector of two values representing the original levels of the grouping variable, in order. These will be relabeled as 'no' and 'yes'. Default: c("Control", "Depression")}
}
\value{
A data frame summarizing phylum-level statistics, including abundance, prevalence, and ASV prevalence per condition and overall.
}
\description{
Computes relative abundance and prevalence of phyla across a binary variable (e.g., condition). Outputs and returns a table with detailed summary statistics.
}
\details{
This function:
\itemize{
\item Converts the input variable (e.g., "Condition") to binary labels "no" and "yes"
\item Calculates phylum-level abundance and prevalence per sample and condition
\item Computes ASV-level prevalence per phylum
\item Outputs a table with summary statistics such as total prevalence, relative prevalence, and mean ASV prevalence
}
}
\examples{
\dontrun{
if(interactive()){
 library(phyloseq)
  # Load example dataset
  data(GlobalPatterns)

  # Create a binary variable in sample_data
  sample_data(GlobalPatterns)$MyCondition <- ifelse(sample_data(GlobalPatterns)$SampleType == "Feces", "Control", "Depression")

  # Run the function summarizing by MyCondition
  result <- getRelAbundancesByPhylumAndVariable(GlobalPatterns, variable="MyCondition", outname="")

  # View the summary table
  head(result)
 }
}
}
\seealso{
\code{\link[dplyr]{summarise}}, \code{\link[dplyr]{mutate}}
\code{\link[tidyr]{spread}}
}
