% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/getRelAbundancesByPhylumAndVariable_beforeAfter.R
\name{getRelAbundancesByPhylumAndVariable_beforeAfter}
\alias{getRelAbundancesByPhylumAndVariable_beforeAfter}
\title{FUNCTION_TITLE}
\usage{
getRelAbundancesByPhylumAndVariable_beforeAfter(
  phobj,
  variable = "Condition",
  outname = "prevalence_by_phylum.tsv"
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{variable}{PARAM_DESCRIPTION, Default: 'Condition'}

\item{outname}{PARAM_DESCRIPTION, Default: 'prevalence_by_phylum.tsv'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{summarise}}, \code{\link[dplyr]{mutate}}
\code{\link[tidyr]{spread}}
}
