% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRelAbundanceTab.R
\name{getRelAbundanceTab}
\alias{getRelAbundanceTab}
\title{FUNCTION_TITLE}
\usage{
getRelAbundanceTab(phobj)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[phyloseq]{otu_table}}, \code{\link[phyloseq]{taxa_sums}}
}
