% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignASVBoxplot_paired.R
\name{getSignASVBoxplot_paired}
\alias{getSignASVBoxplot_paired}
\title{FUNCTION_TITLE}
\usage{
getSignASVBoxplot_paired(
  resdf_annot,
  raw_counts,
  outdir = "./",
  name = "sign_asvs_boxplot.pdf",
  mode = "PROP",
  max_n = 10
)
}
\arguments{
\item{resdf_annot}{PARAM_DESCRIPTION}

\item{raw_counts}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION, Default: './'}

\item{name}{PARAM_DESCRIPTION, Default: 'sign_asvs_boxplot.pdf'}

\item{mode}{PARAM_DESCRIPTION, Default: 'PROP'}

\item{max_n}{PARAM_DESCRIPTION, Default: 10}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{mutate}}
\code{\link[tidyr]{gather}}
}
