% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignif.R
\name{getSignif}
\alias{getSignif}
\title{FUNCTION_TITLE}
\usage{
getSignif(vector)
}
\arguments{
\item{vector}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
