% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignif2.R
\name{getSignif2}
\alias{getSignif2}
\title{FUNCTION_TITLE}
\usage{
getSignif2(
  vector,
  limits = c(0.05, 0.01, 0.001),
  legends = c("", "*", "**", "***")
)
}
\arguments{
\item{vector}{PARAM_DESCRIPTION}

\item{limits}{PARAM_DESCRIPTION, Default: c(0.05, 0.01, 0.001)}

\item{legends}{PARAM_DESCRIPTION, Default: c("", "\emph{", "\strong{", "}}")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
