% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryTablesDeseq.R
\name{getSummaryTablesDeseq}
\alias{getSummaryTablesDeseq}
\title{FUNCTION_TITLE}
\usage{
getSummaryTablesDeseq(res, opt)
}
\arguments{
\item{res}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}
}
