% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTableFromConfmatrices.R
\name{getTableFromConfmatrices}
\alias{getTableFromConfmatrices}
\title{FUNCTION_TITLE}
\usage{
getTableFromConfmatrices(modlist)
}
\arguments{
\item{modlist}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}
}
