% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ensemble_votes.R
\name{make_ensemble_votes}
\alias{make_ensemble_votes}
\title{FUNCTION_TITLE}
\usage{
make_ensemble_votes(
  datasc,
  levs,
  modlist,
  model_res,
  param = "Kappa_l1out",
  min_val = 0,
  prop = TRUE,
  only_1_knn = FALSE
)
}
\arguments{
\item{datasc}{PARAM_DESCRIPTION}

\item{levs}{PARAM_DESCRIPTION}

\item{modlist}{PARAM_DESCRIPTION}

\item{model_res}{PARAM_DESCRIPTION}

\item{param}{PARAM_DESCRIPTION, Default: 'Kappa_l1out'}

\item{min_val}{PARAM_DESCRIPTION, Default: 0}

\item{prop}{PARAM_DESCRIPTION, Default: TRUE}

\item{only_1_knn}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{filter}}
\code{\link[purrr]{map}}
}
