% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_IPAQ_Boxplot.R
\name{make_IPAQ_Boxplot}
\alias{make_IPAQ_Boxplot}
\title{FUNCTION_TITLE}
\usage{
make_IPAQ_Boxplot(
  divtab,
  var = "IPAQ_act_fisica",
  test2show = "wilcox.test",
  alpha_indices,
  outdir,
  name,
  signif_levels = c(`***` = 0.001, `**` = 0.01, `*` = 0.05, ns = 1.1),
  w = 10,
  h = 4,
  correct_pvalues = TRUE
)
}
\arguments{
\item{divtab}{PARAM_DESCRIPTION}

\item{var}{PARAM_DESCRIPTION, Default: 'IPAQ_act_fisica'}

\item{test2show}{PARAM_DESCRIPTION, Default: 'wilcox.test'}

\item{alpha_indices}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}

\item{signif_levels}{PARAM_DESCRIPTION, Default: c(\verb{***} = 0.001, \verb{**} = 0.01, \code{*} = 0.05, ns = 1.1)}

\item{w}{PARAM_DESCRIPTION, Default: 10}

\item{h}{PARAM_DESCRIPTION, Default: 4}

\item{correct_pvalues}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate}}
\code{\link[ggsignif]{stat_signif}}
}
