% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_svm_l1o.R
\name{make_svm_l1o}
\alias{make_svm_l1o}
\title{FUNCTION_TITLE}
\usage{
make_svm_l1o(
  datasc,
  levs,
  varnames,
  kernel = "linear",
  SEED = 123,
  folds = c(),
  do_smote = FALSE,
  smote_params = list(K = 5, dup_size = 2),
  balance_classes = TRUE
)
}
\arguments{
\item{datasc}{PARAM_DESCRIPTION}

\item{levs}{PARAM_DESCRIPTION}

\item{varnames}{PARAM_DESCRIPTION}

\item{kernel}{PARAM_DESCRIPTION, Default: 'linear'}

\item{SEED}{PARAM_DESCRIPTION, Default: 123}

\item{folds}{PARAM_DESCRIPTION, Default: c()}

\item{do_smote}{PARAM_DESCRIPTION, Default: FALSE}

\item{smote_params}{PARAM_DESCRIPTION, Default: list(K = 5, dup_size = 2)}

\item{balance_classes}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}
\code{\link[e1071]{svm}}
}
