% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAllPCAs.R
\name{makeAllPCAs}
\alias{makeAllPCAs}
\title{FUNCTION_TITLE}
\usage{
makeAllPCAs(phobj, counts_df, genes, vars2pca, opt, name = "PCAs")
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{counts_df}{PARAM_DESCRIPTION}

\item{genes}{PARAM_DESCRIPTION}

\item{vars2pca}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION, Default: 'PCAs'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
