% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBarplotFunctional.R
\name{makeBarplotFunctional}
\alias{makeBarplotFunctional}
\title{FUNCTION_TITLE}
\usage{
makeBarplotFunctional(
  tab,
  plim = 0.001,
  name,
  outdir,
  include_longnames = FALSE,
  w = 12,
  h = 12
)
}
\arguments{
\item{tab}{PARAM_DESCRIPTION}

\item{plim}{PARAM_DESCRIPTION, Default: 0.001}

\item{name}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION}

\item{include_longnames}{PARAM_DESCRIPTION, Default: FALSE}

\item{w}{PARAM_DESCRIPTION, Default: 12}

\item{h}{PARAM_DESCRIPTION, Default: 12}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{mutate}}
}
