% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBetaDispTests.R
\name{makeBetaDispTests}
\alias{makeBetaDispTests}
\title{FUNCTION_TITLE}
\usage{
makeBetaDispTests(
  phobj,
  dist_method = "bray",
  exclude_vars = c("sampleID"),
  outname
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{dist_method}{PARAM_DESCRIPTION, Default: 'bray'}

\item{exclude_vars}{PARAM_DESCRIPTION, Default: c("sampleID")}

\item{outname}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[phyloseq]{distance}}
}
