% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeConstrainedOrdination.R
\name{makeConstrainedOrdination}
\alias{makeConstrainedOrdination}
\title{FUNCTION_TITLE}
\usage{
makeConstrainedOrdination(
  phobj,
  variables = c("Condition"),
  dist_type = "bray",
  color_by = ""
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{variables}{PARAM_DESCRIPTION, Default: c("Condition")}

\item{dist_type}{PARAM_DESCRIPTION, Default: 'bray'}

\item{color_by}{PARAM_DESCRIPTION, Default: ''}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[phyloseq]{distance}}
\code{\link[vegan]{scores}}
}
