% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeConstrainedOrdinationSingleVar.R
\name{makeConstrainedOrdinationSingleVar}
\alias{makeConstrainedOrdinationSingleVar}
\title{FUNCTION_TITLE}
\usage{
makeConstrainedOrdinationSingleVar(
  phobj,
  variables = c("Condition"),
  dist_type = "bray"
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{variables}{PARAM_DESCRIPTION, Default: c("Condition")}

\item{dist_type}{PARAM_DESCRIPTION, Default: 'bray'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[phyloseq]{distance}}
\code{\link[vegan]{scores}}
}
