% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeConstrainedOrdinationSingleVar.R
\name{makeConstrainedOrdinationSingleVar}
\alias{makeConstrainedOrdinationSingleVar}
\title{makeConstrainedOrdinationSingleVar}
\usage{
makeConstrainedOrdinationSingleVar(
  phobj,
  variables = c("Condition"),
  dist_type = "bray"
)
}
\arguments{
\item{phobj}{phyloseq object}

\item{variables}{Vector of variables to include in the ordination formula, Default: c("Condition")}

\item{dist_type}{Distance to use, Default: 'bray'}
}
\value{
A ggplot object
}
\description{
Uses the phyloseq::ordinate method with the CAP method
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[phyloseq]{distance}}
\code{\link[vegan]{scores}}
}
