% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeContingencyPlot.R
\name{makeContingencyPlot}
\alias{makeContingencyPlot}
\title{FUNCTION_TITLE}
\usage{
makeContingencyPlot(df, var1, var2, outdir, name, w = 8, h = 6)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{var1}{PARAM_DESCRIPTION}

\item{var2}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}

\item{w}{PARAM_DESCRIPTION, Default: 8}

\item{h}{PARAM_DESCRIPTION, Default: 6}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
