% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCorrelationHeatmap.R
\name{makeCorrelationHeatmap}
\alias{makeCorrelationHeatmap}
\title{FUNCTION_TITLE}
\usage{
makeCorrelationHeatmap(
  mat1,
  mat2,
  cormethod = "pearson",
  pval = 0.05,
  select_asvs = c(),
  outdir = "",
  name = "corrHeatmap",
  clust = T,
  order_rows = c(),
  order_cols = c()
)
}
\arguments{
\item{mat1}{PARAM_DESCRIPTION}

\item{mat2}{PARAM_DESCRIPTION}

\item{cormethod}{PARAM_DESCRIPTION, Default: 'pearson'}

\item{pval}{PARAM_DESCRIPTION, Default: 0.05}

\item{select_asvs}{PARAM_DESCRIPTION, Default: c()}

\item{outdir}{PARAM_DESCRIPTION, Default: ''}

\item{name}{PARAM_DESCRIPTION, Default: 'corrHeatmap'}

\item{clust}{PARAM_DESCRIPTION, Default: T}

\item{order_rows}{PARAM_DESCRIPTION, Default: c()}

\item{order_cols}{PARAM_DESCRIPTION, Default: c()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tidyr]{spread}}
}
