% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFullMediationAnalysisIMC.R
\name{makeFullMediationAnalysisIMC}
\alias{makeFullMediationAnalysisIMC}
\title{FUNCTION_TITLE}
\usage{
makeFullMediationAnalysisIMC(
  vstdf,
  df_all,
  opt,
  getVarsFunction,
  mediator_name = "IMC",
  y_name = "Condition_bin",
  plim = 0.05,
  plim_plot = 0.05,
  name = "analysis_IMC_separateModel_vjust",
  wnet = 14,
  hnet = 12,
  wbars = 8,
  hbars = 10,
  wbars2 = 10,
  hbars2 = 12,
  use_color_scale = FALSE,
  fix_barplot_limits = FALSE,
  custom_colors = NULL,
  make_boxplots = TRUE,
  list2merge = NULL,
  make_power_test = FALSE,
  min_power = 0.9
)
}
\arguments{
\item{vstdf}{PARAM_DESCRIPTION}

\item{df_all}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{getVarsFunction}{PARAM_DESCRIPTION}

\item{mediator_name}{PARAM_DESCRIPTION, Default: 'IMC'}

\item{y_name}{PARAM_DESCRIPTION, Default: 'Condition_bin'}

\item{plim}{PARAM_DESCRIPTION, Default: 0.05}

\item{plim_plot}{PARAM_DESCRIPTION, Default: 0.05}

\item{name}{PARAM_DESCRIPTION, Default: 'analysis_IMC_separateModel_vjust'}

\item{wnet}{PARAM_DESCRIPTION, Default: 14}

\item{hnet}{PARAM_DESCRIPTION, Default: 12}

\item{wbars}{PARAM_DESCRIPTION, Default: 8}

\item{hbars}{PARAM_DESCRIPTION, Default: 10}

\item{wbars2}{PARAM_DESCRIPTION, Default: 10}

\item{hbars2}{PARAM_DESCRIPTION, Default: 12}

\item{use_color_scale}{PARAM_DESCRIPTION, Default: FALSE}

\item{fix_barplot_limits}{PARAM_DESCRIPTION, Default: FALSE}

\item{custom_colors}{PARAM_DESCRIPTION, Default: NULL}

\item{make_boxplots}{PARAM_DESCRIPTION, Default: TRUE}

\item{list2merge}{PARAM_DESCRIPTION, Default: NULL}

\item{make_power_test}{PARAM_DESCRIPTION, Default: FALSE}

\item{min_power}{PARAM_DESCRIPTION, Default: 0.9}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[assertthat]{assert_that}}
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{summarise_all}}
\code{\link[tidyr]{unite}}
}
