% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatmap.R
\name{makeHeatmap}
\alias{makeHeatmap}
\title{FUNCTION_TITLE}
\usage{
makeHeatmap(
  resdf,
  dds,
  df2plot,
  variable = "condition",
  opt,
  name = "heatmap.pdf",
  logscale = FALSE,
  ptype = "padj",
  w = 5,
  h = 4,
  trim_values = FALSE,
  italics_rownames = TRUE,
  taxalist = c()
)
}
\arguments{
\item{resdf}{PARAM_DESCRIPTION}

\item{dds}{PARAM_DESCRIPTION}

\item{df2plot}{PARAM_DESCRIPTION}

\item{variable}{PARAM_DESCRIPTION, Default: 'condition'}

\item{opt}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION, Default: 'heatmap.pdf'}

\item{logscale}{PARAM_DESCRIPTION, Default: FALSE}

\item{ptype}{PARAM_DESCRIPTION, Default: 'padj'}

\item{w}{PARAM_DESCRIPTION, Default: 5}

\item{h}{PARAM_DESCRIPTION, Default: 4}

\item{trim_values}{PARAM_DESCRIPTION, Default: FALSE}

\item{italics_rownames}{PARAM_DESCRIPTION, Default: TRUE}

\item{taxalist}{PARAM_DESCRIPTION, Default: c()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}
}
