% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatmapFunctional.R
\name{makeHeatmapFunctional}
\alias{makeHeatmapFunctional}
\title{FUNCTION_TITLE}
\usage{
makeHeatmapFunctional(
  resdf,
  met2use,
  df2plot,
  variable = "condition",
  opt,
  name = "heatmap.pdf",
  logscale = FALSE,
  ptype = "padj",
  w = 20,
  h = 14
)
}
\arguments{
\item{resdf}{PARAM_DESCRIPTION}

\item{met2use}{PARAM_DESCRIPTION}

\item{df2plot}{PARAM_DESCRIPTION}

\item{variable}{PARAM_DESCRIPTION, Default: 'condition'}

\item{opt}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION, Default: 'heatmap.pdf'}

\item{logscale}{PARAM_DESCRIPTION, Default: FALSE}

\item{ptype}{PARAM_DESCRIPTION, Default: 'padj'}

\item{w}{PARAM_DESCRIPTION, Default: 20}

\item{h}{PARAM_DESCRIPTION, Default: 14}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}
}
