% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeKnn_l1o.R
\name{makeKnn_l1o}
\alias{makeKnn_l1o}
\title{FUNCTION_TITLE}
\usage{
makeKnn_l1o(
  datasc,
  levs,
  varnames,
  different_ks = c(1, 3, 5, 7, 9, 11, 13),
  folds = c(),
  do_smote = FALSE,
  smote_params = list(K = 5, dup_size = "balance")
)
}
\arguments{
\item{datasc}{PARAM_DESCRIPTION}

\item{levs}{PARAM_DESCRIPTION}

\item{varnames}{PARAM_DESCRIPTION}

\item{different_ks}{PARAM_DESCRIPTION, Default: c(1, 3, 5, 7, 9, 11, 13)}

\item{folds}{PARAM_DESCRIPTION, Default: c()}

\item{do_smote}{PARAM_DESCRIPTION, Default: FALSE}

\item{smote_params}{PARAM_DESCRIPTION, Default: list(K = 5, dup_size = "balance")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}
}
