% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLinearModelsSingleVariable.R
\name{makeLinearModelsSingleVariable}
\alias{makeLinearModelsSingleVariable}
\title{FUNCTION_TITLE}
\usage{
makeLinearModelsSingleVariable(
  divtab,
  interestvar,
  extravars,
  alphaindices = c("Observed", "Chao1", "Shannon", "InvSimpson"),
  combos = 1:3,
  outdir = "",
  name = "linearmodels"
)
}
\arguments{
\item{divtab}{PARAM_DESCRIPTION}

\item{interestvar}{PARAM_DESCRIPTION}

\item{extravars}{PARAM_DESCRIPTION}

\item{alphaindices}{PARAM_DESCRIPTION, Default: c("Observed", "Chao1", "Shannon", "InvSimpson")}

\item{combos}{PARAM_DESCRIPTION, Default: 1:3}

\item{outdir}{PARAM_DESCRIPTION, Default: ''}

\item{name}{PARAM_DESCRIPTION, Default: 'linearmodels'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[base]{subset}}
\code{\link[dplyr]{mutate}}
}
