% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLinePlotComparingPhobjs.R
\name{makeLinePlotComparingPhobjs}
\alias{makeLinePlotComparingPhobjs}
\title{FUNCTION_TITLE}
\usage{
makeLinePlotComparingPhobjs(
  all_model_results,
  opt,
  models_name1 = "padj_taxa_res",
  models_name2 = "praw_taxa_res"
)
}
\arguments{
\item{all_model_results}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{models_name1}{PARAM_DESCRIPTION, Default: 'padj_taxa_res'}

\item{models_name2}{PARAM_DESCRIPTION, Default: 'praw_taxa_res'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{select}}
\code{\link[ggpubr]{theme_pubr}}
}
