% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMediationSimpleLavaan.R
\name{makeMediationSimpleLavaan}
\alias{makeMediationSimpleLavaan}
\title{FUNCTION_TITLE}
\usage{
makeMediationSimpleLavaan(df, xname, yname, medname)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{xname}{PARAM_DESCRIPTION}

\item{yname}{PARAM_DESCRIPTION}

\item{medname}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{filter}}
\code{\link[lavaan]{sem}}
}
