% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNaiveBayes_l1o.R
\name{makeNaiveBayes_l1o}
\alias{makeNaiveBayes_l1o}
\title{FUNCTION_TITLE}
\usage{
makeNaiveBayes_l1o(
  datasc,
  levs,
  varnames,
  SEED = 123,
  folds = c(),
  do_smote = FALSE,
  smote_params = list(K = 5, dup_size = "balance")
)
}
\arguments{
\item{datasc}{PARAM_DESCRIPTION}

\item{levs}{PARAM_DESCRIPTION}

\item{varnames}{PARAM_DESCRIPTION}

\item{SEED}{PARAM_DESCRIPTION, Default: 123}

\item{folds}{PARAM_DESCRIPTION, Default: c()}

\item{do_smote}{PARAM_DESCRIPTION, Default: FALSE}

\item{smote_params}{PARAM_DESCRIPTION, Default: list(K = 5, dup_size = "balance")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}
}
