% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePCBarplot.R
\name{makePCBarplot}
\alias{makePCBarplot}
\title{FUNCTION_TITLE}
\usage{
makePCBarplot(
  phname,
  all_model_results,
  pcBoxplots,
  daa_all,
  opt,
  get_pcnames_from = "padj_taxa_res",
  pca_name = "padj_taxa_pcas",
  varname = "Condition",
  w = 8,
  h = 14
)
}
\arguments{
\item{phname}{PARAM_DESCRIPTION}

\item{all_model_results}{PARAM_DESCRIPTION}

\item{pcBoxplots}{PARAM_DESCRIPTION}

\item{daa_all}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{get_pcnames_from}{PARAM_DESCRIPTION, Default: 'padj_taxa_res'}

\item{pca_name}{PARAM_DESCRIPTION, Default: 'padj_taxa_pcas'}

\item{varname}{PARAM_DESCRIPTION, Default: 'Condition'}

\item{w}{PARAM_DESCRIPTION, Default: 8}

\item{h}{PARAM_DESCRIPTION, Default: 14}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{arrange}}
\code{\link[tidyr]{gather}}
}
