% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePCsBoxplot.R
\name{makePCsBoxplot}
\alias{makePCsBoxplot}
\title{FUNCTION_TITLE}
\usage{
makePCsBoxplot(
  phname,
  all_model_results,
  opt,
  get_pcnames_from = "padj_taxa_res",
  pca_name = "padj_taxa_pcas",
  varname = "Condition",
  w = 4,
  h = 6
)
}
\arguments{
\item{phname}{PARAM_DESCRIPTION}

\item{all_model_results}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{get_pcnames_from}{PARAM_DESCRIPTION, Default: 'padj_taxa_res'}

\item{pca_name}{PARAM_DESCRIPTION, Default: 'padj_taxa_pcas'}

\item{varname}{PARAM_DESCRIPTION, Default: 'Condition'}

\item{w}{PARAM_DESCRIPTION, Default: 4}

\item{h}{PARAM_DESCRIPTION, Default: 6}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{summarise}}
\code{\link[tidyr]{gather}}, \code{\link[tidyr]{spread}}
\code{\link[ggsignif]{stat_signif}}
}
