% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePermanova.R
\name{makePermanova}
\alias{makePermanova}
\title{FUNCTION_TITLE}
\usage{
makePermanova(
  phobj,
  dist_method = "bray",
  seed = 123,
  exclude_vars = c("sampleID"),
  outname = "permanovas.tsv",
  disp_permutations = 1000
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{dist_method}{PARAM_DESCRIPTION, Default: 'bray'}

\item{seed}{PARAM_DESCRIPTION, Default: 123}

\item{exclude_vars}{PARAM_DESCRIPTION, Default: c("sampleID")}

\item{outname}{PARAM_DESCRIPTION, Default: 'permanovas.tsv'}

\item{disp_permutations}{PARAM_DESCRIPTION, Default: 1000}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[phyloseq]{distance}}, \code{\link[phyloseq]{prune_samples}}
\code{\link[dplyr]{arrange}}
}
