% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePermanovaSeveralFactors.R
\name{makePermanovaSeveralFactors}
\alias{makePermanovaSeveralFactors}
\title{FUNCTION_TITLE}
\usage{
makePermanovaSeveralFactors(
  phobj,
  dist_method = "bray",
  seed = 123,
  modlist = c(),
  outname = "permanovas_mult.RData"
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{dist_method}{PARAM_DESCRIPTION, Default: 'bray'}

\item{seed}{PARAM_DESCRIPTION, Default: 123}

\item{modlist}{PARAM_DESCRIPTION, Default: c()}

\item{outname}{PARAM_DESCRIPTION, Default: 'permanovas_mult.RData'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[phyloseq]{distance}}
}
