% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRarefactionCurve.R
\name{makeRarefactionCurve}
\alias{makeRarefactionCurve}
\title{FUNCTION_TITLE}
\usage{
makeRarefactionCurve(
  phyloseq_rawdata,
  opt,
  add_hlines = TRUE,
  name = "rarecurve"
)
}
\arguments{
\item{phyloseq_rawdata}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{add_hlines}{PARAM_DESCRIPTION, Default: TRUE}

\item{name}{PARAM_DESCRIPTION, Default: 'rarecurve'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{summarise}}
}
