% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSankeyPlot.R
\name{makeSankeyPlot}
\alias{makeSankeyPlot}
\title{FUNCTION_TITLE}
\usage{
makeSankeyPlot(nodelist, linklist, name, fname, outdir, fontsize = 18)
}
\arguments{
\item{nodelist}{PARAM_DESCRIPTION}

\item{linklist}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}

\item{fname}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION}

\item{fontsize}{PARAM_DESCRIPTION, Default: 18}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
\code{\link[htmlwidgets]{saveWidget}}
}
