% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameProcesses.R
\name{nameProcesses}
\alias{nameProcesses}
\title{FUNCTION_TITLE}
\usage{
nameProcesses(tab, namedtab)
}
\arguments{
\item{tab}{PARAM_DESCRIPTION}

\item{namedtab}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
