% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAbundanceFullPipeline.R
\name{plotAbundanceFullPipeline}
\alias{plotAbundanceFullPipeline}
\title{FUNCTION_TITLE}
\usage{
plotAbundanceFullPipeline(
  phobj,
  interestvar,
  outdir,
  phname,
  levs,
  tops = c(5, 10, 15, 20)
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{interestvar}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION}

\item{phname}{PARAM_DESCRIPTION}

\item{levs}{PARAM_DESCRIPTION}

\item{tops}{PARAM_DESCRIPTION, Default: c(5, 10, 15, 20)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
