% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAllModelPredictions.R
\name{plotAllModelPredictions}
\alias{plotAllModelPredictions}
\title{FUNCTION_TITLE}
\usage{
plotAllModelPredictions(
  phname,
  all_model_results,
  opt,
  get_pcnames_from = "padj_taxa_res",
  pca_name = "padj_taxa_pcas",
  varname = "Condition",
  w = 16,
  h = 16,
  pred_mode = "l1o"
)
}
\arguments{
\item{phname}{PARAM_DESCRIPTION}

\item{all_model_results}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{get_pcnames_from}{PARAM_DESCRIPTION, Default: 'padj_taxa_res'}

\item{pca_name}{PARAM_DESCRIPTION, Default: 'padj_taxa_pcas'}

\item{varname}{PARAM_DESCRIPTION, Default: 'Condition'}

\item{w}{PARAM_DESCRIPTION, Default: 16}

\item{h}{PARAM_DESCRIPTION, Default: 16}

\item{pred_mode}{PARAM_DESCRIPTION, Default: 'l1o'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[cowplot]{plot_grid}}
}
