% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{FUNCTION_TITLE}
\usage{
plotPCA(counts, design, genes, plotvar = "Condition")
}
\arguments{
\item{counts}{PARAM_DESCRIPTION}

\item{design}{PARAM_DESCRIPTION}

\item{genes}{PARAM_DESCRIPTION}

\item{plotvar}{PARAM_DESCRIPTION, Default: 'Condition'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}}
\code{\link[cowplot]{plot_grid}}
}
