% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhylumBoxplots.R
\name{plotPhylumBoxplots}
\alias{plotPhylumBoxplots}
\title{FUNCTION_TITLE}
\usage{
plotPhylumBoxplots(
  phobj,
  var = "Condition",
  outname = "phylumBarplot.pdf",
  height = 10,
  width = 8,
  paired = F
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{var}{PARAM_DESCRIPTION, Default: 'Condition'}

\item{outname}{PARAM_DESCRIPTION, Default: 'phylumBarplot.pdf'}

\item{height}{PARAM_DESCRIPTION, Default: 10}

\item{width}{PARAM_DESCRIPTION, Default: 8}

\item{paired}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[phyloseq]{tax_glom}}, \code{\link[phyloseq]{taxa_names}}, \code{\link[phyloseq]{tax_table}}, \code{\link[phyloseq]{psmelt}}
}
