% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPrediction.R
\name{plotPrediction}
\alias{plotPrediction}
\title{FUNCTION_TITLE}
\usage{
plotPrediction(
  phname,
  mod2plot,
  all_model_results,
  opt,
  get_pcnames_from = "padj_taxa_res",
  pca_name = "padj_taxa_pcas",
  varname = "Condition",
  pred_mode = "l1o",
  w = 6,
  h = 4
)
}
\arguments{
\item{phname}{PARAM_DESCRIPTION}

\item{mod2plot}{PARAM_DESCRIPTION}

\item{all_model_results}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{get_pcnames_from}{PARAM_DESCRIPTION, Default: 'padj_taxa_res'}

\item{pca_name}{PARAM_DESCRIPTION, Default: 'padj_taxa_pcas'}

\item{varname}{PARAM_DESCRIPTION, Default: 'Condition'}

\item{pred_mode}{PARAM_DESCRIPTION, Default: 'l1o'}

\item{w}{PARAM_DESCRIPTION, Default: 6}

\item{h}{PARAM_DESCRIPTION, Default: 4}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
\code{\link[caret]{confusionMatrix}}
}
