% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPrevalenceVsAbundance.R
\name{plotPrevalenceVsAbundance}
\alias{plotPrevalenceVsAbundance}
\title{FUNCTION_TITLE}
\usage{
plotPrevalenceVsAbundance(
  phobj,
  outname = "phylumBarplot.pdf",
  height = 10,
  width = 12
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{outname}{PARAM_DESCRIPTION, Default: 'phylumBarplot.pdf'}

\item{height}{PARAM_DESCRIPTION, Default: 10}

\item{width}{PARAM_DESCRIPTION, Default: 12}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
