% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegressionsASV_vs_vars2.R
\name{plotRegressionsASV_vs_vars2}
\alias{plotRegressionsASV_vs_vars2}
\title{FUNCTION_TITLE}
\usage{
plotRegressionsASV_vs_vars2(
  df3,
  corrdf_annot,
  vars2cor,
  groupvar = "Psoriasis",
  group_levels = c("no", "yes"),
  top_n = 10,
  groupnames = c("Control", "Psoriasis"),
  ylabel = "log(pg/mL)",
  outdir = "",
  name = "regr.pdf",
  w = 7,
  h = 14,
  opt = list()
)
}
\arguments{
\item{df3}{PARAM_DESCRIPTION}

\item{corrdf_annot}{PARAM_DESCRIPTION}

\item{vars2cor}{PARAM_DESCRIPTION}

\item{groupvar}{PARAM_DESCRIPTION, Default: 'Psoriasis'}

\item{group_levels}{PARAM_DESCRIPTION, Default: c("no", "yes")}

\item{top_n}{PARAM_DESCRIPTION, Default: 10}

\item{groupnames}{PARAM_DESCRIPTION, Default: c("Control", "Psoriasis")}

\item{ylabel}{PARAM_DESCRIPTION, Default: 'log(pg/mL)'}

\item{outdir}{PARAM_DESCRIPTION, Default: ''}

\item{name}{PARAM_DESCRIPTION, Default: 'regr.pdf'}

\item{w}{PARAM_DESCRIPTION, Default: 7}

\item{h}{PARAM_DESCRIPTION, Default: 14}

\item{opt}{PARAM_DESCRIPTION, Default: list()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{desc}}
\code{\link[cowplot]{plot_grid}}
}
