% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRelativeAbnBars_Fantaxtic.R
\name{plotRelativeAbnBars_Fantaxtic}
\alias{plotRelativeAbnBars_Fantaxtic}
\title{FUNCTION_TITLE}
\usage{
plotRelativeAbnBars_Fantaxtic(
  phobj,
  variable = "Condition",
  topn = 15,
  tax_level = "Genus",
  outname = "GenusBarplotFx.pdf",
  height = 7,
  width = 12,
  wespalette = "AsteroidCity1"
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{variable}{PARAM_DESCRIPTION, Default: 'Condition'}

\item{topn}{PARAM_DESCRIPTION, Default: 15}

\item{tax_level}{PARAM_DESCRIPTION, Default: 'Genus'}

\item{outname}{PARAM_DESCRIPTION, Default: 'GenusBarplotFx.pdf'}

\item{height}{PARAM_DESCRIPTION, Default: 7}

\item{width}{PARAM_DESCRIPTION, Default: 12}

\item{wespalette}{PARAM_DESCRIPTION, Default: 'AsteroidCity1'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
