% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRelativeAbnBarsGenus.R
\name{plotRelativeAbnBarsGenus}
\alias{plotRelativeAbnBarsGenus}
\title{FUNCTION_TITLE}
\usage{
plotRelativeAbnBarsGenus(
  phobj,
  variable = "Condition",
  topn = 15,
  outname = "phylumBarplot.pdf",
  height = 8,
  width = 12,
  ocluster = T,
  oldlevs = c("Control", "Depression"),
  wespalette = "Darjeeling2"
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{variable}{PARAM_DESCRIPTION, Default: 'Condition'}

\item{topn}{PARAM_DESCRIPTION, Default: 15}

\item{outname}{PARAM_DESCRIPTION, Default: 'phylumBarplot.pdf'}

\item{height}{PARAM_DESCRIPTION, Default: 8}

\item{width}{PARAM_DESCRIPTION, Default: 12}

\item{ocluster}{PARAM_DESCRIPTION, Default: T}

\item{oldlevs}{PARAM_DESCRIPTION, Default: c("Control", "Depression")}

\item{wespalette}{PARAM_DESCRIPTION, Default: 'Darjeeling2'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{summarise_all}}, \code{\link[dplyr]{mutate_all}}
\code{\link[tidyr]{gather}}
}
