% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRelativeAbnBarsSpecies_ColByGenus.R
\name{plotRelativeAbnBarsSpecies_ColByGenus}
\alias{plotRelativeAbnBarsSpecies_ColByGenus}
\title{FUNCTION_TITLE}
\usage{
plotRelativeAbnBarsSpecies_ColByGenus(
  phobj,
  variable = "Condition",
  topn = 15,
  outname = "SpeciescolByGenusBarplot.pdf",
  height = 8,
  width = 12,
  docluster = T
)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{variable}{PARAM_DESCRIPTION, Default: 'Condition'}

\item{topn}{PARAM_DESCRIPTION, Default: 15}

\item{outname}{PARAM_DESCRIPTION, Default: 'SpeciescolByGenusBarplot.pdf'}

\item{height}{PARAM_DESCRIPTION, Default: 8}

\item{width}{PARAM_DESCRIPTION, Default: 12}

\item{docluster}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[phyloseq]{transform_sample_counts}}
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate_all}}
\code{\link[tidyr]{gather}}
}
