% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSVM.R
\name{plotSVM}
\alias{plotSVM}
\title{FUNCTION_TITLE}
\usage{
plotSVM(modelo_svm, datasc, varnames, opt, name)
}
\arguments{
\item{modelo_svm}{PARAM_DESCRIPTION}

\item{datasc}{PARAM_DESCRIPTION}

\item{varnames}{PARAM_DESCRIPTION}

\item{opt}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}
}
