% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSankeyPlot.R
\name{prepareSankeyPlot}
\alias{prepareSankeyPlot}
\title{FUNCTION_TITLE}
\usage{
prepareSankeyPlot(
  daa_proc_all,
  daa_by_sp_all,
  quant_by_proc,
  metadf,
  plim = 0.01,
  plim_sp = 0.01,
  include_others = FALSE,
  include_longnames = FALSE,
  others_name = "Aggr. taxa",
  case_name = "Depression",
  control_name = "Control",
  outdir = "~/",
  name = "test",
  ntop = 0
)
}
\arguments{
\item{daa_proc_all}{PARAM_DESCRIPTION}

\item{daa_by_sp_all}{PARAM_DESCRIPTION}

\item{quant_by_proc}{PARAM_DESCRIPTION}

\item{metadf}{PARAM_DESCRIPTION}

\item{plim}{PARAM_DESCRIPTION, Default: 0.01}

\item{plim_sp}{PARAM_DESCRIPTION, Default: 0.01}

\item{include_others}{PARAM_DESCRIPTION, Default: FALSE}

\item{include_longnames}{PARAM_DESCRIPTION, Default: FALSE}

\item{others_name}{PARAM_DESCRIPTION, Default: 'Aggr. taxa'}

\item{case_name}{PARAM_DESCRIPTION, Default: 'Depression'}

\item{control_name}{PARAM_DESCRIPTION, Default: 'Control'}

\item{outdir}{PARAM_DESCRIPTION, Default: '~/'}

\item{name}{PARAM_DESCRIPTION, Default: 'test'}

\item{ntop}{PARAM_DESCRIPTION, Default: 0}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{summarise_all}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{arrange}}
\code{\link[scales]{pal_gradient_n}}
}
