% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctionalMatrix.R
\name{readFunctionalMatrix}
\alias{readFunctionalMatrix}
\title{FUNCTION_TITLE}
\usage{
readFunctionalMatrix(opt, fname)
}
\arguments{
\item{opt}{PARAM_DESCRIPTION}

\item{fname}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
