% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDiversityDifferences.R
\name{testDiversityDifferences}
\alias{testDiversityDifferences}
\title{FUNCTION_TITLE}
\usage{
testDiversityDifferences(
  divtab,
  vars,
  groupvars,
  outdir,
  name = "alpha_diversity"
)
}
\arguments{
\item{divtab}{PARAM_DESCRIPTION}

\item{vars}{PARAM_DESCRIPTION}

\item{groupvars}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION, Default: 'alpha_diversity'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
