% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDiversityWithQuantVars.R
\name{testDiversityWithQuantVars}
\alias{testDiversityWithQuantVars}
\title{FUNCTION_TITLE}
\usage{
testDiversityWithQuantVars(
  divtab,
  vars,
  qvars,
  outdir,
  name = "AlphaDiv_quant_vars_regressions"
)
}
\arguments{
\item{divtab}{PARAM_DESCRIPTION}

\item{vars}{PARAM_DESCRIPTION}

\item{qvars}{PARAM_DESCRIPTION}

\item{outdir}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION, Default: 'AlphaDiv_quant_vars_regressions'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate_all}}, \code{\link[dplyr]{mutate}}
\code{\link[broom]{reexports}}
}
