% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testHMP.R
\name{testHMP}
\alias{testHMP}
\title{FUNCTION_TITLE}
\usage{
testHMP(phobj, seed = 123)
}
\arguments{
\item{phobj}{PARAM_DESCRIPTION}

\item{seed}{PARAM_DESCRIPTION, Default: 123}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[phyloseq]{tax_glom}}, \code{\link[phyloseq]{subset_samples}}, \code{\link[phyloseq]{otu_table}}
\code{\link[HMP]{Xdc.sevsample}}
}
